% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{add_column_types_header_line}
\alias{add_column_types_header_line}
\title{add_column_types_header_line}
\usage{
add_column_types_header_line(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Overwrite any previous column type lines? Logical}
}
\value{
Nothing.
}
\description{
One-off function to insert a `Column types` header line into input files.
}
