% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{add_comments}
\alias{add_comments}
\title{add_comments}
\usage{
add_comments(x, comments)
}
\arguments{
\item{x}{An object}

\item{comments}{A character vector of comments}
}
\value{
\code{x} with comments appended to any existing comments.
}
\description{
Add character comments to a data system object. Comments are written out
with the data when the file is saved.
}

