% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{add_dsflags}
\alias{add_dsflags}
\title{add_dsflag}
\usage{
add_dsflags(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{One or more flags (that can be coerced to character)}
}
\value{
\code{x} with flags appended to any existing flags
}
\description{
Add character flags to a data system object. Flags are used internally, and in some
cases (for testing data) are written out with the data when the file is saved.
}

