% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{add_logit_tables_xml}
\alias{add_logit_tables_xml}
\title{Add a table which has definitions with logit.type to be converted to XML}
\usage{
add_logit_tables_xml(dot, data, header, base_logit_header = header)
}
\arguments{
\item{dot}{The current state of the pipeline started from \code{create_xml}}

\item{data}{The tibble of data to add to the conversion}

\item{header}{The header tag to can be looked up in the header file to
convert \code{data}}

\item{base_logit_header}{The base header tag to use for the generated logit type
tables such that \code{paste(base_logit_header, logit.type, sep="_")} corresponds
to the appropriate model interface header.  Note by default this value is the
same as \code{header} as often this is the case but certainly not always.}
}
\value{
A "data structure" to hold the various parts needed to run the model
interface CSV to XML conversion.
}
\description{
Such tables need to get partitioned into a series of tables due to limitations
in the way that the Model Interfaces processes information.  In particular all
XML tag names have to be specified in the header and can not be set dynamically
as is the intention with the logit.type.
}
\note{
For documentation of the options for logit choice functions in GCAM, see
\url{http://jgcri.github.io/gcam-doc/choice.html}.
}
\author{
Pralit Patel
}
