% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{add_node_equiv_xml}
\alias{add_node_equiv_xml}
\title{add_node_equiv_xml}
\usage{
add_node_equiv_xml(dot, equiv_class)
}
\arguments{
\item{dot}{The current state of the pipeline started from \code{create_xml}}

\item{equiv_class}{A name of the equivalence class to add.  This could be any
key in the list \code{XML_NODE_EQUIV}.}
}
\value{
A "data structure" to hold the various parts needed to run the model
interface CSV to XML conversion.
}
\description{
Add a table to an XML pipeline that instructs the ModelInterface to treat
tags in the same class as the same when enabled.  Thus not requiring multiple
headers for instance to read in a share-weight into a technology
intermittent-technology or tranTechnology.  A user taking advantage of this
feature would then read this table early in the XML pipeline.
}
\author{
Pralit Patel
}
