% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{add_node_leaf_names}
\alias{add_node_leaf_names}
\title{add_node_leaf_names}
\usage{
add_node_leaf_names(data, nesting_table, leaf_name, LT_name = "Land_Type",
  LN1 = "LandNode1", LN2 = NA, LN3 = NA, LN4 = NA, append_GLU = TRUE)
}
\arguments{
\item{data}{data}

\item{nesting_table}{nesting_table}

\item{leaf_name}{leaf_name}

\item{LT_name}{LT_name}

\item{LN1}{LN1}

\item{LN2}{LN2}

\item{LN3}{LN3}

\item{LN4}{LN4}

\item{append_GLU}{append_GLU}
}
\value{
return
}
\description{
Match in the node and leaf names from a land nesting table
}
