% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{add_reference}
\alias{add_reference}
\title{add_reference}
\usage{
add_reference(x, reference)
}
\arguments{
\item{x}{An object}

\item{reference}{Reference for object (character)}
}
\value{
\code{x} with units appended to any existing references
}
\description{
Add character units to a data system object. Units are written out
with the data when the file is saved.
}

