% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{add_title}
\alias{add_title}
\title{add_title}
\usage{
add_title(x, title)
}
\arguments{
\item{x}{An object}

\item{title}{Title of object (character)}
}
\value{
\code{x} with units appended to any existing comments.
}
\description{
Add character units to a data system object. Units are written out
with the data when the file is saved.
}

