% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_xml_data}
\alias{add_xml_data}
\title{add_xml_data}
\usage{
add_xml_data(dot, data, header)
}
\arguments{
\item{dot}{The current state of the pipeline started from \code{create_xml}.}

\item{data}{The tibble of data to add to the conversion.}

\item{header}{The header tag to can be looked up in the header file to
convert \code{data}.}
}
\value{
A "data structure" to hold the various parts needed to run the model
interface CSV to XML conversion.
}
\description{
Add a table to include for conversion to XML.  We need the tibble to convert
and a header tag which can be looked up in the header file to convert the
tibble.  This method is meant to be included in a pipeline between calls of
\code{\link{create_xml}} and \code{\link{run_xml_conversion}}.
}

