% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-aglu.R
\name{aglu_LA100.FAO_downscale_ctry_makedata}
\alias{aglu_LA100.FAO_downscale_ctry_makedata}
\title{aglu_LA100.FAO_downscale_ctry_makedata}
\usage{
aglu_LA100.FAO_downscale_ctry_makedata(all_data)
}
\arguments{
\item{all_data}{A named list, holding all data system products so far}
}
\value{
A named list with all aglu data.
}
\description{
aglu_LA100.FAO_downscale_ctry_makedata
}

