% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{append_GLU}
\alias{append_GLU}
\title{append_GLU}
\usage{
append_GLU(data, var1 = "LandNode1", var2 = NA, var3 = NA, var4 = NA,
  var5 = NA)
}
\arguments{
\item{data}{data}

\item{var1}{var1}

\item{var2}{var2}

\item{var3}{var3}

\item{var4}{var4}

\item{var5}{var5}
}
\value{
return
}
\description{
Append GLU to all specified variables
}
