% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{approx_fun}
\alias{approx_fun}
\title{approx_fun}
\usage{
approx_fun(year, value, rule = 1)
}
\arguments{
\item{year}{Numeric year, in a melted tibble or data frame}

\item{value}{Numeric value to interpolate}

\item{rule}{Rule to use; see \code{\link{approx}} and details}
}
\value{
Interpolated values.
}
\description{
\code{\link{approx}} for use in a dplyr pipeline.
}
\details{
This was \code{gcam_interp} in the original data system.
}
\examples{
df <- data.frame(year = 1:5, value = c(1, 2, NA, 4, 5))
approx_fun(df$year, df$value, rule = 2)
}

