% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atest.R
\name{atest_makedata}
\alias{atest_makedata}
\title{atest_makedata}
\usage{
atest_makedata(all_data)
}
\arguments{
\item{all_data}{A named list, holding all data system products so far}
}
\value{
A named list with all available test data
}
\description{
atest_makedata
}

