% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunk_dependencies}
\alias{chunk_dependencies}
\title{chunk_dependencies}
\usage{
chunk_dependencies(chunks = find_chunks()$name)
}
\arguments{
\item{chunks}{A character vector of chunks names}
}
\value{
A tibble with columns 'name' (chunk name) and 'input' (name of data)
}
\description{
chunk_dependencies
}

