% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunk_outputs}
\alias{chunk_outputs}
\title{chunk_outputs}
\usage{
chunk_outputs(chunks = find_chunks()$name)
}
\arguments{
\item{chunks}{A character vector of chunks names}
}
\value{
A tibble with columns 'name' (chunk name) and 'output' (name of data)
}
\description{
List all chunk outputs.
}
