% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{chunk_readylist}
\alias{chunk_readylist}
\title{chunk_readylist}
\usage{
chunk_readylist()
}
\value{
Returns a list of disabled chunks: their lines of code, inputs and available inputs, dependencies
}
\description{
chunk_readylist
}

