% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{cmp_xml_files}
\alias{cmp_xml_files}
\title{Run the XML diff utility on two XML files}
\usage{
cmp_xml_files(fleft, fright, raw = FALSE)
}
\arguments{
\item{fleft}{The first file to compare}

\item{fright}{The second file to compare}

\item{raw}{Flag: if true run in detail mode}
}
\description{
Comparing two XML files turns out to be more easily done in python, so this
function runs the python comparison code.  The standard mode checks the
return code and returns TRUE if the files are equivalent, or FALSE if they
are not.  If the files are not equivalent, the differences found will be
written to the console.  Comparing two files that aren't equivalent will also
raise a warning due to the nonzero status code returned when differences are
found.
}
\details{
The 'raw' mode is intended to be used in tests.  It returns the stdout
string generated by the comparison code.  If the files are equivalent this
string will be empty.  If not, then it will contain the diagnostic output,
and it will have a 'status' attribute giving the return code
}
