% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_datamap_from_cache}
\alias{create_datamap_from_cache}
\title{create_datamap_from_cache}
\usage{
create_datamap_from_cache(gcamdata_plan, ...)
}
\arguments{
\item{gcamdata_plan}{The drake plan that generated the outputs}

\item{...}{Parameters to pass directly to drake::readd to allow
users to specify a non-default cache for instance.}
}
\value{
The full GCAM data map
}
\description{
Re-creates the GCAM data map by pulling the data out of cache and processing
its metadata.  We need to know the plan to be able to identify which data to
to pull and the chunk names that generated them.
}
