% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aglu.R
\name{downscale_FAO_country}
\alias{downscale_FAO_country}
\title{downscale_FAO_country}
\usage{
downscale_FAO_country(data, country_name, dissolution_year,
  item_name = "item", element_name = "element",
  years = AGLU_HISTORICAL_YEARS)
}
\arguments{
\item{data}{Data to downscale}

\item{country_name}{Country name, pre-dissolution}

\item{dissolution_year}{Year of country dissolution}

\item{item_name}{Item column name}

\item{element_name}{Element column name}

\item{years}{Years to operate on}
}
\value{
Downscaled data.
}
\description{
Helper function to downscale the countries that separated into
multiple modern countries (e.g. USSR).
}

