% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-socioeconomics.R
\name{downscale_Maddison_country}
\alias{downscale_Maddison_country}
\title{downscale_Maddison_country}
\usage{
downscale_Maddison_country(data, downscale_country_name, socioeconomics_ctry,
  available_year)
}
\arguments{
\item{data}{data}

\item{downscale_country_name}{country_name}

\item{socioeconomics_ctry}{socioeconomics_ctry}

\item{available_year}{available_year}
}
\value{
something
}
\description{
function to downscale the countries that separated into multiple modern countries (e.g. USSR).
}

