% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstrace.R
\name{dstrace_plot}
\alias{dstrace_plot}
\title{Plot a trace}
\usage{
dstrace_plot(object_name, tracelist, upstream, downstream, ...)
}
\arguments{
\item{object_name}{Name of original object being traced}

\item{tracelist}{Record of the trace, a tibble}

\item{upstream}{Looking upstream? Logical}

\item{downstream}{Looking downstream? Logical}

\item{...}{Extra arguments passed on to \code{\link{plot}}}
}
\value{
Adjacency matrix, invisible
}
\description{
Plot a trace
}
