% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-energy.R
\name{energy_LA113.MSW}
\alias{energy_LA113.MSW}
\title{energy_LA113.MSW}
\usage{
energy_LA113.MSW(all_data)
}
\arguments{
\item{all_data}{A named list, holding all energy system products so far}
}
\value{
A named list with all energy data.
}
\description{
energy_LA113.MSW
}

