% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_csv_file}
\alias{find_csv_file}
\title{find_csv_file}
\usage{
find_csv_file(filename)
}
\arguments{
\item{filename}{Bare (no extension) filename to find}
}
\value{
full name of file
}
\description{
Find an internal, i.e. included with the package, data file.
}

