% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{find_multiple_mutates}
\alias{find_multiple_mutates}
\title{find_multiple_mutates}
\usage{
find_multiple_mutates(path = "R/")
}
\arguments{
\item{path}{Path to look in}
}
\value{
Nothing.
}
\description{
find_multiple_mutates
}
\details{
Internal/admin function to find files with sequential 'mutate' calls,
because mutate(x) %>% mutate(y) incurs a BIG performance penalty versus
mutate(x, y) so we try to not have the former case.
}
\note{
This will find some false positives (right now in zchunk_L124.nonco2_unmgd_R_S_T_Y.R
and zchunk_L203.demand_input.R) as well as false negatives (e.g. mutates separated by comments)
}
\author{
BBL November 2018
}
