% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-gcam-usa.R
\name{gcam-usa_LA100.Socioeconomics_makedata}
\alias{gcam-usa_LA100.Socioeconomics_makedata}
\title{gcam-usa_LA100.Socioeconomics_makedata}
\usage{
"gcam-usa_LA100.Socioeconomics_makedata"(all_data)
}
\arguments{
\item{all_data}{A named list, holding all data system products so far}
}
\value{
A named list with all gcam-usa data.
}
\description{
gcam-usa_LA100.Socioeconomics_makedata
}

