% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{generate_level2_data_names}
\alias{generate_level2_data_names}
\title{generate_level2_data_names}
\usage{
generate_level2_data_names()
}
\value{
A list object where [[header]] <- character vector of column names
corresponding to header.
}
\description{
Generate the list of "level 2 data names" or really a list of column orderings
keyed by the ModelInterface header so that we can ensure tables being sent to be
converted to XML by the ModelInterface have their columns arranged in the order
the ModelInterface is expecting them.
}
\author{
Pralit Patel
}
