% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{get_data}
\alias{get_data}
\title{getdata}
\usage{
get_data(all_data, name)
}
\arguments{
\item{all_data}{Data structure}

\item{name}{Name of data to return}
}
\value{
Data object (currently, a tibble or data frame).
}
\description{
This function returns a tibble (currently) in \code{all_data},
abstracting away the mechanism for accessing it from the chunks.
Throws an error if data is not present.
}
