% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{get_data_}
\alias{get_data_}
\title{getdata v2}
\usage{
get_data_(name, all_data)
}
\arguments{
\item{name}{Name of data to return}

\item{all_data}{Data structure}
}
\value{
Data object (currently, a tibble or data frame). If the object was marked
\code{NA} in the data store, indicating an optional input that was not found,
a \code{NULL} is returned.
}
\description{
Same as get_data, but with arguments swapped, allowing
apply functions to be used across a vector of
}
