% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{get_logit_fn_tables}
\alias{get_logit_fn_tables}
\title{get_logit_fn_tables}
\usage{
get_logit_fn_tables(data, names, base_header, include_equiv_table = TRUE,
  write_all_regions = FALSE, default_logit_type = gcam.LOGIT_TYPES[1], ...)
}
\arguments{
\item{data}{Data to partition by logit.type, tibble}

\item{names}{Column names to use out of data, character}

\item{base_header}{Base header that is used for the logit type tables which will get
pasted with what the logit.type for each table, character}

\item{include_equiv_table}{Include EQUIV_TABLE as well? Logical}

\item{write_all_regions}{Write each table to all regions? Logical}

\item{default_logit_type}{Default logit function to use if the user did not specify one, character}

\item{...}{Other parameters to pass to \code{\link{write_to_all_regions}}}
}
\value{
The tables TODO NEEED BETTER DOCUMENTATION
}
\description{
Generate a list of tables that sets the appropriate discrete choice function to use.
}
\details{
The data has to be partitioned into multiple tables, one for each logit.type. Each element of
the returned list has two variables: 1) the header for the table, and 2) the data for the table.
If requested, an additional \code{EQUIV_TABLE} table will be added so that the tables that contain the logit
exponents do not themselves have to know what logit.type they are using and thus do not need to be
partitioned.
}
