% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{get_ssp4_regions}
\alias{get_ssp4_regions}
\title{get_ssp4_regions}
\usage{
get_ssp4_regions(pcGDP, reg_names, income_group)
}
\arguments{
\item{pcGDP}{A tibble with per capita GDP estimates}

\item{reg_names}{A tibble with columns \code{GCAM_region_ID} and \code{region}}

\item{income_group}{A string indicating which region group (low, medium, high)}
}
\value{
A list of region names belonging to the specified income group.
}
\description{
Get regions for different income groups in SSP4
}
