% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{get_water_inputs_for_mapping}
\alias{get_water_inputs_for_mapping}
\title{get_water_inputs_for_mapping}
\usage{
get_water_inputs_for_mapping(x, water_mapping)
}
\arguments{
\item{x}{A tibble containing a \code{water.sector} and \code{water_type}
(and \code{GLU} if the water sector is irrigation) to determine the new input names}

\item{water_mapping}{A tibble with a mapping between \code{water.sector} and  \code{supplysector}}
}
\value{
A vector of names of form supplysector_watertype or supplysector_GLU_watertype.
}
\description{
Get the appropriate minicam.energy.input name to use in the GCAM supplysector.
}
\details{
Get the appropriate minicam.energy.input name to use in the GCAM supplysector
by looking up using a mapping to the water.sector and water_type. The minicam.energy.input
name to use will have to be some water mapping sector for water_types that are "mapped".
}
