% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inputs_of}
\alias{inputs_of}
\title{inputs_of}
\usage{
inputs_of(chunks)
}
\arguments{
\item{chunks}{Names of chunks, character}
}
\value{
Character vector of inputs.
}
\description{
Convenience function for getting the inputs of one or more chunks
}
