% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{is_data_list}
\alias{is_data_list}
\title{is_data_list}
\usage{
is_data_list(data_list)
}
\arguments{
\item{data_list}{List of data frames or other objects}
}
\value{
TRUE or FALSE.
}
\description{
Check whether an object is a valid list (collection) of data objects.
}
\details{
Currently a data_list is just a list.
}

