% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\docType{data}
\name{LEVEL2_DATA_NAMES}
\alias{LEVEL2_DATA_NAMES}
\title{A list of column orderings keyed by the ModelInterface header so that we can
ensure tables being sent to be converted to XML by the ModelInterface have their
columns arranged in the order the ModelInterface is expecting them.}
\format{An object of class \code{list} of length 321.}
\usage{
LEVEL2_DATA_NAMES
}
\description{
A list of column orderings keyed by the ModelInterface header so that we can
ensure tables being sent to be converted to XML by the ModelInterface have their
columns arranged in the order the ModelInterface is expecting them.
}
\author{
Pralit Patel
}
\keyword{datasets}
