% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_csv}
\alias{load_csv}
\title{load_csv}
\usage{
load_csv(filename, ...)
}
\arguments{
\item{filename}{Name of file to load}

\item{...}{Any other parameter to pass to \code{readr::read_csv}}
}
\value{
data frame from file
}
\description{
Load an internal, i.e. included with the package, csv (or csv.gz) data file
}
\details{
The data frame read in is marked as an input, not one that has
been computed, via \code{\link{add_dscomments}}.
}

