% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_csv_file}
\alias{load_csv_file}
\title{load_csv_file}
\usage{
load_csv_file(filename, quiet = FALSE, ...)
}
\arguments{
\item{filename}{Filename to load}

\item{quiet}{Logical - suppress messages?}

\item{...}{Any other parameter to pass to \code{readr::read_csv}}
}
\value{
A list of data frames (tibbles).
}
\description{
Load one or more internal, i.e. included with the package, csv (or csv.gz) data files.
}
\details{
The data frames read in are marked as inputs, not ones that have
been computed, via \code{\link{add_dscomments}}.
}

