% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime.R
\name{modeltime}
\alias{modeltime}
\title{modeltime}
\usage{
modeltime()
}
\value{
A named list with all modeltime components
}
\description{
Construct the \code{modeltime} data structures
}
\examples{
modeltime()
}
\author{
BBL
}

