% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime.R
\name{modeltime_makedata}
\alias{modeltime_makedata}
\title{modeltime_makedata}
\usage{
modeltime_makedata(all_data)
}
\arguments{
\item{all_data}{A named list, holding all data system products so far}
}
\value{
A named list with all available modeltime data
}
\description{
modeltime_makedata
}

