% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_ag_prodchange_ssp2_IRR_MGMT_xml.R
\name{module_aglu_batch_ag_prodchange_ssp2_IRR_MGMT_xml}
\alias{module_aglu_batch_ag_prodchange_ssp2_IRR_MGMT_xml}
\title{module_aglu_batch_ag_prodchange_ssp2_IRR_MGMT_xml}
\usage{
module_aglu_batch_ag_prodchange_ssp2_IRR_MGMT_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{ag_prodchange_ssp2_IRR_MGMT.xml}. The corresponding file in the
original data system was \code{batch_ag_prodchange_ssp2_IRR_MGMT.xml.R} (aglu XML).
}
\description{
Construct XML data structure for \code{ag_prodchange_ssp2_IRR_MGMT.xml}.
}
