% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_land_input_2.xml.R
\name{module_aglu_batch_land_input_2.xml_DISABLED}
\alias{module_aglu_batch_land_input_2.xml_DISABLED}
\title{module_aglu_batch_land_input_2.xml}
\usage{
module_aglu_batch_land_input_2.xml_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{land_input_2.xml}. The corresponding file in the
original data system was \code{batch_land_input_2.xml.R} (aglu XML).
}
\description{
Construct XML data structure for \code{land_input_2.xml}.
}
