% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L201.ag_For_Past_bio_input.R
\name{module_aglu_L201.ag_For_Past_bio_input_DISABLED}
\alias{module_aglu_L201.ag_For_Past_bio_input_DISABLED}
\title{module_aglu_L201.ag_For_Past_bio_input}
\usage{
module_aglu_L201.ag_For_Past_bio_input_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L201.AgSectorLogitTables[[ curr_table ]]$data}, \code{L201.AgSupplySector}, \code{L201.AgSubsectorLogitTables[[ curr_table ]]$data}, \code{L201.AgSupplySubsector}, \code{L201.AgProduction_ag}, \code{L201.AgProduction_For}, \code{L201.AgProduction_Past}, \code{L201.AgHAtoCL}, \code{L201.AgYield_bio_grass}, \code{L201.AgYield_bio_tree}. The corresponding file in the
original data system was \code{L201.ag_For_Past_bio_input.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
