% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2012.ag_For_Past_bio_input_irr_mgmt.R
\name{module_aglu_L2012.ag_For_Past_bio_input_irr_mgmt}
\alias{module_aglu_L2012.ag_For_Past_bio_input_irr_mgmt}
\title{module_aglu_L2012.ag_For_Past_bio_input_irr_mgmt}
\usage{
module_aglu_L2012.ag_For_Past_bio_input_irr_mgmt(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2012.AgSupplySector}, \code{L2012.AgSupplySubsector}, \code{L2012.AgProduction_ag_irr_mgmt}, \code{L2012.AgProduction_For}, \code{L2012.AgProduction_Past}, \code{L2012.AgHAtoCL_irr_mgmt}, \code{L2012.AgYield_bio_ref}. The corresponding file in the
original data system was \code{L2012.ag_For_Past_bio_input_irr_mgmt.R} (aglu level2).
}
\description{
Builds agriculture, forest, pasture and biomass production inputs for all technologies.
}
\details{
This chunk specifies the input tables for agriculture, forest, pasture and biomass supply sectors and subsectors,
agricultural commodity production and harvest area to cropland by technologies, forest and pasture production,
and biomass grass and tree crops yield by technologies.
}
\author{
RC August 2017
}
