% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L204.resbio_input.R
\name{module_aglu_L204.resbio_input_DISABLED}
\alias{module_aglu_L204.resbio_input_DISABLED}
\title{module_aglu_L204.resbio_input}
\usage{
module_aglu_L204.resbio_input_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L204.AgResBio_For}, \code{L204.GlobalResBio_Mill}, \code{L204.AgResBio_ag}, \code{L204.AgResBioCurve_For}, \code{L204.StubResBioCurve_Mill}, \code{L204.AgResBioCurve_ag}. The corresponding file in the
original data system was \code{L204.resbio_input.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
