% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2041.resbio_input_irr.R
\name{module_aglu_L2041.resbio_input_irr_DISABLED}
\alias{module_aglu_L2041.resbio_input_irr_DISABLED}
\title{module_aglu_L2041.resbio_input_irr}
\usage{
module_aglu_L2041.resbio_input_irr_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2041.AgResBio_For}, \code{L2041.GlobalResBio_Mill}, \code{L2041.AgResBio_ag_irr}, \code{L2041.AgResBioCurve_For}, \code{L2041.StubResBioCurve_Mill}, \code{L2041.AgResBioCurve_ag_irr}. The corresponding file in the
original data system was \code{L2041.resbio_input_irr.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
