% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2042.resbio_input_irr_mgmt.R
\name{module_aglu_L2042.resbio_input_irr_mgmt_DISABLED}
\alias{module_aglu_L2042.resbio_input_irr_mgmt_DISABLED}
\title{module_aglu_L2042.resbio_input_irr_mgmt}
\usage{
module_aglu_L2042.resbio_input_irr_mgmt_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2042.AgResBio_For}, \code{L2042.AgResBioCurve_For}, \code{L2042.GlobalResBio_Mill}, \code{L2042.StubResBioCurve_Mill}, \code{L2042.AgResBio_ag_irr_mgmt}, \code{L2042.AgResBioCurve_ag_irr_mgmt}. The corresponding file in the
original data system was \code{L2042.resbio_input_irr_mgmt.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
