% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L205.ag_prodchange_cost_input.R
\name{module_aglu_L205.ag_prodchange_cost_input_DISABLED}
\alias{module_aglu_L205.ag_prodchange_cost_input_DISABLED}
\title{module_aglu_L205.ag_prodchange_cost_input}
\usage{
module_aglu_L205.ag_prodchange_cost_input_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L205.AgProdChange_ag_ref}, \code{L205.AgProdChange_bio_ref}, \code{L205.AgCost_ag}, \code{L205.AgCost_bio}, \code{L205.AgCost_Past}, \code{L205.AgCost_For}, \code{L205.AgProdChange_high}, \code{L205.AgProdChange_ag_ref}, \code{L205.AgProdChange_low}, \code{L205.AgProdChange_ssp4}, \code{L205.AgProdChange_high}. The corresponding file in the
original data system was \code{L205.ag_prodchange_cost_input.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
