% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L206.ag_Fert.R
\name{module_aglu_L206.ag_Fert_DISABLED}
\alias{module_aglu_L206.ag_Fert_DISABLED}
\title{module_aglu_L206.ag_Fert}
\usage{
module_aglu_L206.ag_Fert_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L206.AgCoef_Fert_ag}, \code{L206.AgCoef_Fert_bio}, \code{L206.AgCost_ag_adj}, \code{L206.AgCost_bio_adj}. The corresponding file in the
original data system was \code{L206.ag_Fert.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
