% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2061.ag_Fert_irr.R
\name{module_aglu_L2061.ag_Fert_irr_DISABLED}
\alias{module_aglu_L2061.ag_Fert_irr_DISABLED}
\title{module_aglu_L2061.ag_Fert_irr}
\usage{
module_aglu_L2061.ag_Fert_irr_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2061.AgCoef_Fert_ag_irr}, \code{L2061.AgCoef_Fert_bio_irr}, \code{L2061.AgCost_ag_irr_adj}, \code{L2061.AgCost_bio_irr_adj}. The corresponding file in the
original data system was \code{L2061.ag_Fert_irr.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
