% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2062.ag_Fert_irr_mgmt.R
\name{module_aglu_L2062.ag_Fert_irr_mgmt}
\alias{module_aglu_L2062.ag_Fert_irr_mgmt}
\title{module_aglu_L2062.ag_Fert_irr_mgmt}
\usage{
module_aglu_L2062.ag_Fert_irr_mgmt(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2062.AgCoef_Fert_ag_irr_mgmt}, \code{L2062.AgCoef_Fert_bio_irr_mgmt}, \code{L2062.AgCost_ag_irr_mgmt_adj}, \code{L2062.AgCost_bio_irr_mgmt_adj}. The corresponding file in the
original data system was \code{L2062.ag_Fert_irr_mgmt.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
