% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2071.ag_water_irr.R
\name{module_aglu_L2071.ag_water_irr_DISABLED}
\alias{module_aglu_L2071.ag_water_irr_DISABLED}
\title{module_aglu_L2071.ag_water_irr}
\usage{
module_aglu_L2071.ag_water_irr_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2071.AgCoef_IrrBphysWater_ag}, \code{L2071.AgCoef_IrrWaterWdraw_ag}, \code{L2071.AgCoef_IrrWaterCons_ag}, \code{L2071.AgCoef_RfdBphysWater_ag}, \code{L2071.AgCoef_BphysWater_bio}, \code{L2071.AgCoef_IrrWaterWdraw_bio}, \code{L2071.AgCoef_IrrWaterCons_bio}. The corresponding file in the
original data system was \code{L2071.ag_water_irr.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
