% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L221.land_input_1.R
\name{module_aglu_L221.land_input_1_DISABLED}
\alias{module_aglu_L221.land_input_1_DISABLED}
\title{module_aglu_L221.land_input_1}
\usage{
module_aglu_L221.land_input_1_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{curr_table$data}, \code{L221.LN0_Logit}, \code{L221.LN0_Land}, \code{L221.LN0_SoilTimeScale}, \code{curr_table$data}, \code{L221.LN1_ValueLogit}, \code{L221.LN1_HistUnmgdAllocation}, \code{L221.LN1_UnmgdAllocation}, \code{L221.LN1_UnmgdCarbon}. The corresponding file in the
original data system was \code{L221.land_input_1.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
