% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L223.land_input_3.R
\name{module_aglu_L223.land_input_3_DISABLED}
\alias{module_aglu_L223.land_input_3_DISABLED}
\title{module_aglu_L223.land_input_3}
\usage{
module_aglu_L223.land_input_3_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{curr_table$data}, \code{L223.LN3_Logit}, \code{L223.LN3_LeafGhostShare}, \code{L223.LN3_LeafIsGhostShareRel}, \code{L223.LN3_HistUnmgdAllocation}, \code{L223.LN3_UnmgdAllocation}, \code{L223.NodeEquiv}, \code{L223.LN3_NoEmissCarbon}, \code{L223.LN3_NodeCarbon}, \code{L223.LN3_HistMgdAllocation_noncrop}, \code{L223.LN3_MgdAllocation_noncrop}, \code{L223.LN3_HistMgdAllocation_crop}, \code{L223.LN3_MgdAllocation_crop}, \code{L223.LN3_HistMgdAllocation_bio}, \code{L223.LN3_MgdAllocation_bio}, \code{L223.LN3_UnmgdCarbon}, \code{L223.LN3_MgdCarbon_noncrop}, \code{L223.LN3_MgdCarbon_crop}, \code{L223.LN3_MgdCarbon_bio}, \code{curr_table$data}, \code{L223.LN1_Logit_prot}, \code{L223.LN3_HistUnmgdAllocation_noprot}, \code{L223.LN3_UnmgdAllocation_noprot}, \code{L223.LN1_HistUnmgdAllocation_prot}, \code{L223.LN1_UnmgdAllocation_prot}, \code{L223.LN1_UnmgdCarbon_prot}. The corresponding file in the
original data system was \code{L223.land_input_3.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
