% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2241.land_input_4_irr.R
\name{module_aglu_L2241.land_input_4_irr_DISABLED}
\alias{module_aglu_L2241.land_input_4_irr_DISABLED}
\title{module_aglu_L2241.land_input_4_irr}
\usage{
module_aglu_L2241.land_input_4_irr_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{curr_table$data}, \code{L2241.LN4_Logit}, \code{L2241.LN4_HistMgdAllocation_crop}, \code{L2241.LN4_MgdAllocation_crop}, \code{L2241.LN4_HistMgdAllocation_bio}, \code{L2241.LN4_MgdAllocation_bio}, \code{L2241.LN4_MgdCarbon_crop}, \code{L2241.LN4_MgdCarbon_bio}, \code{L2241.LN4_LeafGhostShare}, \code{L2241.LN4_NodeGhostShare}, \code{L2241.LN4_NodeIsGhostShareRel}. The corresponding file in the
original data system was \code{L2241.land_input_4_irr.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
