% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2242.land_input_4_irr_mgmt.R
\name{module_aglu_L2242.land_input_4_irr_mgmt}
\alias{module_aglu_L2242.land_input_4_irr_mgmt}
\title{module_aglu_L2242.land_input_4_irr_mgmt}
\usage{
module_aglu_L2242.land_input_4_irr_mgmt(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2242.LN4_Logit}, \code{L2242.LN4_NodeGhostShare}, \code{L2242.LN4_NodeIsGhostShareRel}. The corresponding file in the
original data system was \code{L2242.land_input_4_irr_mgmt.R} (aglu level2).
}
\description{
This chunk generates files of the fourth land nest by region.
}
\details{
This chunk generates the logit exponent of the fourth land nest by region, the ghost node share for bionenergy node in future years,
and specifies whether the bionenergy ghost node share is relative.
}
\author{
RC August 2017
}
