% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2242.land_input_4_irr_mgmt.R
\name{module_aglu_L2242.land_input_4_irr_mgmt_DISABLED}
\alias{module_aglu_L2242.land_input_4_irr_mgmt_DISABLED}
\title{module_aglu_L2242.land_input_4_irr_mgmt}
\usage{
module_aglu_L2242.land_input_4_irr_mgmt_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2241.LN4_LogitTables[[ curr_table_name ]]}, \code{L2242.LN4_Logit}, \code{L2242.LN4_NodeGhostShare}, \code{L2242.LN4_NodeIsGhostShareRel}. The corresponding file in the
original data system was \code{L2242.land_input_4_irr_mgmt.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
