% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L241.trade_input.R
\name{module_aglu_L241.trade_input_DISABLED}
\alias{module_aglu_L241.trade_input_DISABLED}
\title{module_aglu_L241.trade_input}
\usage{
module_aglu_L241.trade_input_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L241.StubAgTradeCoeff_food}, \code{L241.StubAgTradeCoeff_nonfood}, \code{L241.StubAgTradeCoeff_feed}, \code{L241.AgProdTech_RES_output}, \code{L241.RES_Market}. The corresponding file in the
original data system was \code{L241.trade_input.R} (aglu level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
