% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA100.0_LDS_preprocessing.R
\name{module_aglu_LA100.0_LDS_preprocessing}
\alias{module_aglu_LA100.0_LDS_preprocessing}
\title{module_aglu_LA100.0_LDS_preprocessing}
\usage{
module_aglu_LA100.0_LDS_preprocessing(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{object}. The corresponding file in the
original data system was \code{LA100.0_LDS_preprocessing.R} (aglu level1).
}
\description{
Read in and process LDS (Land Data System) files.
}
\details{
Read in the various LDS datasets; regularize their column names and
GLU (Geographic Land Unit) data; change Taiwan ISO to that of mainland China; make LDS_ag_HA_ha and
LDS_ag_prod_t tables consistent. See Di Vittorio, A., P. Kyle, and W. Collins. 2016.
What are the effects of Agro-Ecological Zones and land use region boundaries on
land resource projection using the Global Change Assessment Model? Environmental
Modelling & Software 85, 246-265. http://dx.doi.org/10.1016/j.envsoft.2016.08.016.
}
\author{
BBL March 2017
}
