% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA100.0_LDS_preprocessing.R
\name{module_aglu_LA100.0_LDS_preprocessing_DISABLED}
\alias{module_aglu_LA100.0_LDS_preprocessing_DISABLED}
\title{module_aglu_LA100.0_LDS_preprocessing}
\usage{
module_aglu_LA100.0_LDS_preprocessing_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{object}. The corresponding file in the
original data system was \code{LA100.0_LDS_preprocessing.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
