% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-aglu.R
\name{module_aglu_LA100.FAO_downscale_ctry}
\alias{module_aglu_LA100.FAO_downscale_ctry}
\title{module_aglu_LA100.FAO_downscale_ctry}
\usage{
module_aglu_LA100.FAO_downscale_ctry(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}.
}
\description{
Construct the \code{gcam-usa} data structures.
}
\author{
BBL
}

