% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA100.IMAGE_downscale_ctry_yr.R
\name{module_aglu_LA100.IMAGE_downscale_ctry_yr_DISABLED}
\alias{module_aglu_LA100.IMAGE_downscale_ctry_yr_DISABLED}
\title{module_aglu_LA100.IMAGE_downscale_ctry_yr}
\usage{
module_aglu_LA100.IMAGE_downscale_ctry_yr_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.IMAGE_an_Feedfrac_ctry_C_Sys_Fd_Y}, \code{L100.IMAGE_an_FeedIO_ctry_C_Sys_Y}, \code{L100.IMAGE_an_Prodmixfrac_ctry_C_Y}. The corresponding file in the
original data system was \code{LA100.IMAGE_downscale_ctry_yr.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
