% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA101.ag_FAO_R_C_Y.R
\name{module_aglu_LA101.ag_FAO_R_C_Y_DISABLED}
\alias{module_aglu_LA101.ag_FAO_R_C_Y_DISABLED}
\title{module_aglu_LA101.ag_FAO_R_C_Y}
\usage{
module_aglu_LA101.ag_FAO_R_C_Y_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L101.ag_Food_Mt_R_C_Y}, \code{L101.ag_Food_Pcal_R_C_Y}, \code{L101.ag_kcalg_R_C_Y}, \code{L101.ag_HA_bm2_R_C_Y}, \code{L101.ag_Prod_Mt_R_C_Y}. The corresponding file in the
original data system was \code{LA101.ag_FAO_R_C_Y.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
