% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA102.ag_LDS_R_C_GLU.R
\name{module_aglu_LA102.ag_LDS_R_C_GLU_DISABLED}
\alias{module_aglu_LA102.ag_LDS_R_C_GLU_DISABLED}
\title{module_aglu_LA102.ag_LDS_R_C_GLU}
\usage{
module_aglu_LA102.ag_LDS_R_C_GLU_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L102.ag_HA_bm2_R_C_GLU}, \code{L102.ag_Prod_Mt_R_C_GLU}. The corresponding file in the
original data system was \code{LA102.ag_LDS_R_C_GLU.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
