% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA103.ag_R_C_Y_GLU.R
\name{module_aglu_LA103.ag_R_C_Y_GLU}
\alias{module_aglu_LA103.ag_R_C_Y_GLU}
\title{module_aglu_LA103.ag_R_C_Y_GLU}
\usage{
module_aglu_LA103.ag_R_C_Y_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L103.ag_Prod_Mt_R_C_Y_GLU}, \code{L103.ag_Prod_Mt_R_C_Y}, \code{L103.ag_HA_bm2_R_C_Y_GLU}, \code{L103.ag_Yield_kgm2_R_C_Y_GLU}. The corresponding file in the
original data system was \code{LA103.ag_R_C_Y_GLU.R} (aglu level1).
}
\description{
Calculate production, harvested area, and yield by region, crop, GLU, and year.
}
\details{
We only have production and harvested area by region and GLU for a single
representative base year (circa 2000), and are using that to downscale regional
production and harvested area in all years. So, if GLU223 accounted for 20% of U.S.
corn production in ~2000, then it accounted for 20% of US corn production in all years.
}
\author{
BBL April 2017
}
