% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA105.an_FAO_R_C_Y.R
\name{module_aglu_LA105.an_FAO_R_C_Y}
\alias{module_aglu_LA105.an_FAO_R_C_Y}
\title{module_aglu_LA105.an_FAO_R_C_Y}
\usage{
module_aglu_LA105.an_FAO_R_C_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L105.an_Food_Mt_R_C_Y}, \code{L105.an_Food_Pcal_R_C_Y}, \code{L105.an_kcalg_R_C_Y}, \code{L105.an_Prod_Mt_R_C_Y}, \code{L105.an_Prod_Mt_ctry_C_Y}. The corresponding file in the
original data system was \code{LA105.an_FAO_R_C_Y.R} (aglu level1).
}
\description{
Aggregates FAO animal products consumption and production data to GCAM region / commodity.
}
\details{
This chunk aggregates FAO animal products food consumption and production data up to GCAM commodities and GCAM regions,
and calculates the average animal products caloric content by GCAM region / commodity / year.
}
\author{
RC June 2017
}
