% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA107.an_IMAGE_R_C_Sys_Fd_Y.R
\name{module_aglu_LA107.an_IMAGE_R_C_Sys_Fd_Y}
\alias{module_aglu_LA107.an_IMAGE_R_C_Sys_Fd_Y}
\title{module_aglu_LA107.an_IMAGE_R_C_Sys_Fd_Y}
\usage{
module_aglu_LA107.an_IMAGE_R_C_Sys_Fd_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L107.an_Prod_Mt_R_C_Sys_Fd_Y}, \code{L107.an_Feed_Mt_R_C_Sys_Fd_Y}, \code{L107.an_FeedIO_R_C_Sys_Fd_Y}. The corresponding file in the
original data system was \code{LA107.an_IMAGE_R_C_Sys_Fd_Y.R} (aglu level1).
}
\description{
Build IMAGE Animal Production, Feed Consumption, and Input-Output Coefficients for each GCAM region and GCAM commodity in every year.
}
\details{
This module builds three separate outputs at the GCAM region - GCAM commodity - Management system - Feed type - Year level of resolution.
First, Animal Production is built at the country-level from IMAGE country-level total production, mixed system fraction, and feed type fraction and aggregated to region.
Second, Feed Consumption is built at the country-level from IMAGE country-level input-output coefficients and animal production information and aggregated to region.
Finally, Feed Input-Output coefficients are calculated from region-level Feed Consumption and Animal Production.
}
\author{
ACS April 2017
}
