% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA107.an_IMAGE_R_C_Sys_Fd_Y.R
\name{module_aglu_LA107.an_IMAGE_R_C_Sys_Fd_Y_DISABLED}
\alias{module_aglu_LA107.an_IMAGE_R_C_Sys_Fd_Y_DISABLED}
\title{module_aglu_LA107.an_IMAGE_R_C_Sys_Fd_Y}
\usage{
module_aglu_LA107.an_IMAGE_R_C_Sys_Fd_Y_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L107.an_Prod_Mt_R_C_Sys_Fd_Y}, \code{L107.an_Feed_Mt_R_C_Sys_Fd_Y}, \code{L107.an_FeedIO_R_C_Sys_Fd_Y}. The corresponding file in the
original data system was \code{LA107.an_IMAGE_R_C_Sys_Fd_Y.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
