% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA108.ag_Feed_R_C_Y.R
\name{module_aglu_LA108.ag_Feed_R_C_Y}
\alias{module_aglu_LA108.ag_Feed_R_C_Y}
\title{module_aglu_LA108.ag_Feed_R_C_Y}
\usage{
module_aglu_LA108.ag_Feed_R_C_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L108.ag_Feed_Mt_R_C_Y}, \code{L108.ag_NetExp_Mt_R_FodderHerb_Y}. The corresponding file in the
original data system was \code{LA108.ag_Feed_R_C_Y.R} (aglu level1).
}
\description{
Compute (1) feed by GCAM commodity, region, and year, and (2) net exports of FodderHerb.
}
\details{
Computes (1) feed by GCAM commodity, region, and year, and (2) net exports of FodderHerb.
The feed calculations combine IMAGE and FAO data together, using FAO to disaggregate IMAGE to GCAM_commodity.
Some adjustments are needed to deal with inconsistencies between IMAGE demand & FAO supply.
Excess supply is mapped to OtherUses. Excess demand is mapped to other sources (Residue, Pasture).
This information is used to calculate net exports of FodderHerb too.
}
\author{
KVC April 2017
}
