% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB110.For_FAO_R_Y.R
\name{module_aglu_LB110.For_FAO_R_Y}
\alias{module_aglu_LB110.For_FAO_R_Y}
\title{module_aglu_LB110.For_FAO_R_Y}
\usage{
module_aglu_LB110.For_FAO_R_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L110.For_ALL_bm3_R_Y}. The corresponding file in the
original data system was \code{LB110.For_FAO_R_Y.R} (aglu level1).
}
\description{
This module builds Forest Production, Net Export (FAO Exports - FAO Imports), and
Consumption  (Production - Net Exports) information for every GCAM region in each year, from
FAO Production, Export, and Import data.
}
\details{
This module builds Forest Production, Net Export (FAO Exports - FAO Imports), and
Consumption  (Production - Net Exports) information for every GCAM region in each year, from
FAO Production, Export, and Import data. FAO Production data is used to scale Net Exports
at the Region level such that Global Production equals Consumption.
}
\author{
MC and ACS March 2017
}
