% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB110.For_FAO_R_Y.R
\name{module_aglu_LB110.For_FAO_R_Y_DISABLED}
\alias{module_aglu_LB110.For_FAO_R_Y_DISABLED}
\title{module_aglu_LB110.For_FAO_R_Y}
\usage{
module_aglu_LB110.For_FAO_R_Y_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L110.For_ALL_bm3_R_Y}. The corresponding file in the
original data system was \code{LB110.For_FAO_R_Y.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
