% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB111.ag_resbio_R_C.R
\name{module_aglu_LB111.ag_resbio_R_C}
\alias{module_aglu_LB111.ag_resbio_R_C}
\title{module_aglu_LB111.ag_resbio_R_C}
\usage{
module_aglu_LB111.ag_resbio_R_C(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L111.ag_resbio_R_C}. The corresponding file in the
original data system was \code{LB111.ag_resbio_R_C.R} (aglu level1).
}
\description{
Calculate the production-weighted parameters of residue biomass by GCAM region and commodity.
}
\details{
This chunk calculates the production-weighted average
residue biomass parameters by GCAM region and commodity.
}
\author{
RC March 2017
}
