% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB111.ag_resbio_R_C.R
\name{module_aglu_LB111.ag_resbio_R_C_DISABLED}
\alias{module_aglu_LB111.ag_resbio_R_C_DISABLED}
\title{module_aglu_LB111.ag_resbio_R_C}
\usage{
module_aglu_LB111.ag_resbio_R_C_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L111.ag_resbio_R_C}. The corresponding file in the
original data system was \code{LB111.ag_resbio_R_C.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
