% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB113.bio_Yield_R_GLU.R
\name{module_aglu_LB113.bio_Yield_R_GLU}
\alias{module_aglu_LB113.bio_Yield_R_GLU}
\title{module_aglu_LB113.bio_Yield_R_GLU}
\usage{
module_aglu_LB113.bio_Yield_R_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L113.ag_bioYield_GJm2_R_GLU}. The corresponding file in the
original data system was \code{LB113.bio_Yield_R_GLU.R} (aglu level1).
}
\description{
Calculate base year bioenergy yields by GCAM region and GLU
}
\details{
Calculate global average yields for each FAO crop in the base year;
calculate each region / zone / crop's comparative yield; compute bioenergy yields as
this region/zone-specific index multiplied by a base yield.
}
\references{
Wullschleger, S.D., E.B. Davis, M.E. Borsuk, C.A. Gunderson, and L.R. Lynd. 2010.
Biomass production in switchgrass across the United States: database description and determinants
of yield. Agronomy Journal 102: 1158-1168. doi:10.2134/agronj2010.0087.
}
\author{
BBL June 2017
}
