% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB113.bio_Yield_R_GLU.R
\name{module_aglu_LB113.bio_Yield_R_GLU_DISABLED}
\alias{module_aglu_LB113.bio_Yield_R_GLU_DISABLED}
\title{module_aglu_LB113.bio_Yield_R_GLU}
\usage{
module_aglu_LB113.bio_Yield_R_GLU_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L113.ag_bioYield_GJm2_R_GLU}. The corresponding file in the
original data system was \code{LB113.bio_Yield_R_GLU.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
