% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB122.LC_R_Cropland_Yh_GLU.R
\name{module_aglu_LB122.LC_R_Cropland_Yh_GLU}
\alias{module_aglu_LB122.LC_R_Cropland_Yh_GLU}
\title{module_aglu_LB122.LC_R_Cropland_Yh_GLU}
\usage{
module_aglu_LB122.LC_R_Cropland_Yh_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L122.ag_HA_to_CropLand_R_Y_GLU}, \code{L122.ag_EcYield_kgm2_R_C_Y_GLU}, \code{L122.LC_bm2_R_OtherArableLand_Yh_GLU}, \code{L122.LC_bm2_R_ExtraCropLand_Yh_GLU}, \code{L122.LC_bm2_R_HarvCropLand_C_Yh_GLU}, \code{L122.LC_bm2_R_HarvCropLand_Yh_GLU}. The corresponding file in the
original data system was \code{LB122.LC_R_Cropland_Yh_GLU.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
