% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB123.LC_R_MgdPastFor_Yh_GLU.R
\name{module_aglu_LB123.LC_R_MgdPastFor_Yh_GLU_DISABLED}
\alias{module_aglu_LB123.LC_R_MgdPastFor_Yh_GLU_DISABLED}
\title{module_aglu_LB123.LC_R_MgdPastFor_Yh_GLU}
\usage{
module_aglu_LB123.LC_R_MgdPastFor_Yh_GLU_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L123.ag_Prod_Mt_R_Past_Y_GLU}, \code{L123.ag_Yield_kgm2_R_Past_Y_GLU}, \code{L123.LC_bm2_R_MgdPast_Yh_GLU}, \code{L123.For_Prod_bm3_R_Y_GLU}, \code{L123.For_Yield_m3m2_R_GLU}, \code{L123.LC_bm2_R_MgdFor_Yh_GLU}. The corresponding file in the
original data system was \code{LB123.LC_R_MgdPastFor_Yh_GLU.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
