% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB124.LC_R_UnMgd_Yh_GLU.R
\name{module_aglu_LB124.LC_R_UnMgd_Yh_GLU}
\alias{module_aglu_LB124.LC_R_UnMgd_Yh_GLU}
\title{module_aglu_LB124.LC_R_UnMgd_Yh_GLU}
\usage{
module_aglu_LB124.LC_R_UnMgd_Yh_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L124.LC_bm2_R_Shrub_Yh_GLU_adj}, \code{L124.LC_bm2_R_Grass_Yh_GLU_adj}, \code{L124.LC_bm2_R_UnMgdPast_Yh_GLU_adj}, \code{L124.LC_bm2_R_UnMgdFor_Yh_GLU_adj}. The corresponding file in the
original data system was \code{LB124.LC_R_UnMgd_Yh_GLU.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
