% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB125.LC_tot.R
\name{module_aglu_LB125.LC_tot_DISABLED}
\alias{module_aglu_LB125.LC_tot_DISABLED}
\title{module_aglu_LB125.LC_tot}
\usage{
module_aglu_LB125.LC_tot_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L125.LC_bm2_R}, \code{L125.LC_bm2_R_GLU}, \code{L125.LC_bm2_R_LT_Yh_GLU}. The corresponding file in the
original data system was \code{LB125.LC_tot.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
