% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB131.LV_R_GLU.R
\name{module_aglu_LB131.LV_R_GLU}
\alias{module_aglu_LB131.LV_R_GLU}
\title{module_aglu_LB131.LV_R_GLU}
\usage{
module_aglu_LB131.LV_R_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L131.LV_USD75_m2_R_GLU}. The corresponding file in the
original data system was \code{LB131.LV_R_GLU.R} (aglu level1).
}
\description{
Compute average value of land in $/m2
}
\details{
Computes average value of land in $/m2, using total land value
and harvested cropland area.
}
\author{
KVC April 2017
}
