% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB133.ag_Costs_USA_C_2005.R
\name{module_aglu_LB133.ag_Costs_USA_C_2005}
\alias{module_aglu_LB133.ag_Costs_USA_C_2005}
\title{module_aglu_LB133.ag_Costs_USA_C_2005}
\usage{
module_aglu_LB133.ag_Costs_USA_C_2005(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L133.ag_Cost_75USDkg_C}. The corresponding file in the
original data system was \code{LB133.ag_Costs_USA_C_2005.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
