% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB141.ag_Fert_IFA_ctry_crop.R
\name{module_aglu_LB141.ag_Fert_IFA_ctry_crop}
\alias{module_aglu_LB141.ag_Fert_IFA_ctry_crop}
\title{module_aglu_LB141.ag_Fert_IFA_ctry_crop}
\usage{
module_aglu_LB141.ag_Fert_IFA_ctry_crop(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L141.ag_Fert_Cons_MtN_ctry_crop}. The corresponding file in the
original data system was \code{LB141.ag_Fert_IFA_ctry_crop.R} (aglu level1).
}
\description{
This chunk reconciles disparate IFA fertilizer consumption data to calculate fertilizer consumption (demand) for each GTAP country/crop.
}
\details{
Multiple harvested area data sources (LDS, FAO) are reconciled and used with bottom-up fertilizer consumption data from
IFA2002 to calculate fertilizer demand for each country and crop.
Top down estimates are calculated using IFA fertilizer data, and the top down estimates are used to fill in missing data from
the bottom up estimates and scale the bottom-up estimates, making the final output.
}
\author{
ACS May 2017
}
