% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB141.ag_Fert_IFA_ctry_crop.R
\name{module_aglu_LB141.ag_Fert_IFA_ctry_crop_DISABLED}
\alias{module_aglu_LB141.ag_Fert_IFA_ctry_crop_DISABLED}
\title{module_aglu_LB141.ag_Fert_IFA_ctry_crop}
\usage{
module_aglu_LB141.ag_Fert_IFA_ctry_crop_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L141.ag_Fert_Cons_MtN_ctry_crop}. The corresponding file in the
original data system was \code{LB141.ag_Fert_IFA_ctry_crop.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
