% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB151.ag_MIRCA_ctry_C_GLU_irr.R
\name{module_aglu_LB151.ag_MIRCA_ctry_C_GLU_irr}
\alias{module_aglu_LB151.ag_MIRCA_ctry_C_GLU_irr}
\title{module_aglu_LB151.ag_MIRCA_ctry_C_GLU_irr}
\usage{
module_aglu_LB151.ag_MIRCA_ctry_C_GLU_irr(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L151.ag_irrHA_ha_ctry_crop}, \code{L151.ag_rfdHA_ha_ctry_crop}, \code{L151.ag_irrProd_t_ctry_crop}, \code{L151.ag_rfdProd_t_ctry_crop}. The corresponding file in the
original data system was \code{LB151.ag_MIRCA_ctry_C_GLU_irr.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
