% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB162.ag_prodchange_R_C_Y_GLU_irr.R
\name{module_aglu_LB162.ag_prodchange_R_C_Y_GLU_irr}
\alias{module_aglu_LB162.ag_prodchange_R_C_Y_GLU_irr}
\title{module_aglu_LB162.ag_prodchange_R_C_Y_GLU_irr}
\usage{
module_aglu_LB162.ag_prodchange_R_C_Y_GLU_irr(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L162.ag_YieldRatio_R_C_Ysy_GLU_irr}, \code{L162.ag_YieldRate_R_C_Y_GLU_irr}, \code{L162.bio_YieldRate_R_Y_GLU_irr}. The corresponding file in the
original data system was \code{LB162.ag_prodchange_R_C_Y_GLU_irr.R} (aglu level1).
}
\description{
This module calculates the first level production/yield change assumptions that are exogenous to GCAM. These rates are calculated for each commodity
at the region-glu-irrigation level in each model year, including the calibration year.
}
\details{
The CROSIT agriculture database is processed and reconciled with LDS and GCAM data system production and harvested area information at the irrigation
level. Yield Ratios are calculated as future year reconciled production / base year reconciled production. The biomass yield ratio in each year is taken to
be the median of all other commodities at the region-glu-irrigation level. The yield ratios are used to calculate annual yield rate assumptions
(Yield Rate(year i) = Yield Ratio(year i) / Yield ratio(year i-1) ). Externally defined default yield rates are used to fill in missing data at the GCAM
region-commodity-glu-irrigation level for all model years.
}
\author{
ACS June 2017
}
