% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB163.bio_Yield_R_GLU_irr.R
\name{module_aglu_LB163.bio_Yield_R_GLU_irr}
\alias{module_aglu_LB163.bio_Yield_R_GLU_irr}
\title{module_aglu_LB163.bio_Yield_R_GLU_irr}
\usage{
module_aglu_LB163.bio_Yield_R_GLU_irr(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L163.ag_irrBioYield_GJm2_R_GLU}, \code{L163.ag_rfdBioYield_GJm2_R_GLU}. The corresponding file in the
original data system was \code{LB163.bio_Yield_R_GLU_irr.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
