% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB165.ag_water_R_C_Y_GLU_irr.R
\name{module_aglu_LB165.ag_water_R_C_Y_GLU_irr}
\alias{module_aglu_LB165.ag_water_R_C_Y_GLU_irr}
\title{module_aglu_LB165.ag_water_R_C_Y_GLU_irr}
\usage{
module_aglu_LB165.ag_water_R_C_Y_GLU_irr(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L165.BlueIrr_m3kg_R_C_GLU}, \code{L165.TotIrr_m3kg_R_C_GLU}, \code{L165.GreenRfd_m3kg_R_C_GLU}, \code{L165.ag_IrrEff_R}. The corresponding file in the
original data system was \code{LB165.ag_water_R_C_Y_GLU_irr.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
