% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB171.LC_R_Cropland_Yh_GLU_irr.R
\name{module_aglu_LB171.LC_R_Cropland_Yh_GLU_irr}
\alias{module_aglu_LB171.LC_R_Cropland_Yh_GLU_irr}
\title{module_aglu_LB171.LC_R_Cropland_Yh_GLU_irr}
\usage{
module_aglu_LB171.LC_R_Cropland_Yh_GLU_irr(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L171.LC_bm2_R_irrHarvCropLand_C_Yh_GLU}, \code{L171.LC_bm2_R_rfdHarvCropLand_C_Yh_GLU}, \code{L171.ag_irrEcYield_kgm2_R_C_Y_GLU}, \code{L171.ag_rfdEcYield_kgm2_R_C_Y_GLU}. The corresponding file in the
original data system was \code{LB171.LC_R_Cropland_Yh_GLU_irr.R} (aglu level1).
}
\description{
Calculate irrigated/rainfed harvested cropland cover and economic yields by GCAM region / commodity / year / GLU.
}
\details{
This chunk downscales total harvested cropland by GCAM region / commodity / year / GLU to irrigated/rainfed according to irrigated/rainfed shares in the base year, and calculates the economic yields as production divided by cropland.
}
\author{
RC May 2017
}
