% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB181.ag_R_C_Y_GLU_irr_mgmt.R
\name{module_aglu_LB181.ag_R_C_Y_GLU_irr_mgmt}
\alias{module_aglu_LB181.ag_R_C_Y_GLU_irr_mgmt}
\title{module_aglu_LB181.ag_R_C_Y_GLU_irr_mgmt}
\usage{
module_aglu_LB181.ag_R_C_Y_GLU_irr_mgmt(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L181.LC_bm2_R_C_Yh_GLU_irr_level}, \code{L181.ag_EcYield_kgm2_R_C_Y_GLU_irr_level}, \code{L181.ag_Prod_Mt_R_C_Y_GLU_irr_level}, \code{L181.YieldMult_R_bio_GLU_irr}, \code{L181.LandShare_R_bio_GLU_irr}. The corresponding file in the
original data system was \code{LB181.ag_R_C_Y_GLU_irr_mgmt.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
