% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atest.R
\name{module_atest}
\alias{module_atest}
\title{module_atest}
\usage{
module_atest(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}
}
\description{
A test module that can't make data until modeltime data is available
}
\examples{
module_atest("MAKE", all_data = NULL)
}
\author{
BBL
}

