% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_no_climate_model_xml.R
\name{module_batch_no_climate_model_xml}
\alias{module_batch_no_climate_model_xml}
\title{module_batch_no_climate_model_xml}
\usage{
module_batch_no_climate_model_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{no_climate_model.xml}, \code{disable_climate_model.xml}.
}
\description{
Construct XML data structure for \code{no_climate_model.xml}.
}
