% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_hector.xml.R
\name{module_climate_xml_batch_hector.xml}
\alias{module_climate_xml_batch_hector.xml}
\title{module_climate_xml_batch_hector.xml}
\usage{
module_climate_xml_batch_hector.xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{hector.xml}. The corresponding file in the
original data system was \code{batch_hector.xml.R} (climate_xml XML).
}
\description{
Construct XML data structure for \code{hector.xml}.
}
