% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_data_EDGAR_gases.R
\name{module_data_EDGAR_gases}
\alias{module_data_EDGAR_gases}
\title{module_data_EDGAR_gases}
\usage{
module_data_EDGAR_gases(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{EDGAR_gases}.
}
\description{
Dedicated data chunk to read \code{EDGAR} csv files.
}
\details{
This reads all the EDGAR csv files and binds them together.
}
\author{
BBL May 2017
}
