% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_data_UCD_transportation.R
\name{module_data_UCD_transportation}
\alias{module_data_UCD_transportation}
\title{module_data_UCD_transportation}
\usage{
module_data_UCD_transportation(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{UCD_trn_data_CORE}, \code{UCD_trn_data_SSP1},
\code{UCD_trn_data_CORE}, \code{UCD_trn_data_CORE}, \code{UCD_trn_data_CORE}.
}
\description{
Dedicated data chunk to read UCD transportation database
}
\details{
The \code{UCD_trn_data_CORE.csv} file suddenly, in line 1403, stops
using trailing commas for blank columns, and then restarts doing so 672 rows later.
This generates a warning in the general purpose \code{\link{load_csv_files}}
that we'd rather not suppress universally.
}
\author{
BBL June 2017
}
