% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_data_UCD_trn.R
\name{module_data_UCD_trn}
\alias{module_data_UCD_trn}
\title{module_data_UCD_trn}
\usage{
module_data_UCD_trn(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{UCD_trn_data}.
}
\description{
Dedicated data chunk to read \code{Maddison_population.csv} file.
}
\details{
The \code{UCD_trn_data} files have blank columns, which
generates a warning message in the general purpose \code{\link{load_csv_files}}
that we'd rather not suppress universally.
}
\author{
RH
}
