% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_all_aglu_emissions_IRR.xml.R
\name{module_emissions_batch_all_aglu_emissions_IRR.xml_DISABLED}
\alias{module_emissions_batch_all_aglu_emissions_IRR.xml_DISABLED}
\title{module_emissions_batch_all_aglu_emissions_IRR.xml}
\usage{
module_emissions_batch_all_aglu_emissions_IRR.xml_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{all_aglu_emissions_IRR.xml}. The corresponding file in the
original data system was \code{batch_all_aglu_emissions_IRR.xml} (emissions XML).
}
\description{
Construct XML data structure for \code{all_aglu_emissions_IRR.xml}.
}
