% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_MACC_TC_SSP.xml.R
\name{module_emissions_batch_MACC_TC_SSP.xml}
\alias{module_emissions_batch_MACC_TC_SSP.xml}
\title{module_emissions_batch_MACC_TC_SSP.xml}
\usage{
module_emissions_batch_MACC_TC_SSP.xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{MACC_TC_SSP1.xml}, \code{MACC_TC_SSP2.xml}, and \code{MACC_TC_SSP5.xml}.
}
\description{
Construct XML data structure for all the \code{MACC_TC_SSP.xml} files.
}
