% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L101.nonghg_en_USA_S_T_Y.R
\name{module_emissions_L101.nonghg_en_USA_S_T_Y}
\alias{module_emissions_L101.nonghg_en_USA_S_T_Y}
\title{module_emissions_L101.nonghg_en_USA_S_T_Y}
\usage{
module_emissions_L101.nonghg_en_USA_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L101.so2_tgej_USA_en_Sepa_F_Yh}, \code{L101.co_tgej_USA_en_Sepa_F_Yh}, \code{L101.nox_tgej_USA_en_Sepa_F_Yh}, \code{L101.voc_tgej_USA_en_Sepa_F_Yh}, \code{L101.nh3_tgej_USA_en_Sepa_F_Yh}, \code{L101.in_EJ_R_en_Si_F_Yh}. The corresponding file in the
original data system was \code{L101.nonghg_en_USA_S_T_Y.R} (emissions level1).
}
\description{
Compute historical emissions factors for energy by GCAM technology, from EPA emissions data and IEA energy balances.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
