% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L105.nh3_an_USA_S_T_Y.R
\name{module_emissions_L105.nh3_an_USA_S_T_Y}
\alias{module_emissions_L105.nh3_an_USA_S_T_Y}
\title{module_emissions_L105.nh3_an_USA_S_T_Y}
\usage{
module_emissions_L105.nh3_an_USA_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L105.nh3_tgmt_USA_an_Yh}. The corresponding file in the
original data system was \code{L105.nh3_an_USA_S_T_Y.R} (emissions level1).
}
\description{
Historical NH3 emissions factors for animals by GCAM technology, computed from EPA emissions data and FAO animal statistics
}
\details{
Historical NH3 emission factors for animal production are calculated from EPA NH3 estimates (1990 - 2002)
and FAO production statistics by GCAM sector and technology.
Emission factors for historical years (1971 - 1989) are estiamted as the emissions factor for 1990.
All historical NH3 Emission factors use US values.
}
\author{
RMH April 2017
}
