% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L111.nonghg_en_R_S_T_Y.R
\name{module_emissions_L111.nonghg_en_R_S_T_Y}
\alias{module_emissions_L111.nonghg_en_R_S_T_Y}
\title{module_emissions_L111.nonghg_en_R_S_T_Y}
\usage{
module_emissions_L111.nonghg_en_R_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L111.nonghg_tg_R_en_S_F_Yh}, \code{L111.nonghg_tgej_R_en_S_F_Yh}. The corresponding file in the
original data system was \code{L111.nonghg_en_R_S_T_Y.R} (emissions level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
This code produces two outputs: non-ghg emission totals and non-ghg emission shares by total emissions.
First, non-ghg gas emissions were combined and grouped by sector and region.  Then, emissions were scaled, and international
shipping & aviation emission data was calculated based on total emission and total emission shares. Then non-ghg emission
totals and shares were calculated by GCAM sector, fuel, technology, and driver type for EDGAR historical years.
}
\author{
CDL August 2017
}
