% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L111.nonghg_en_R_S_T_Y.R
\name{module_emissions_L111.nonghg_en_R_S_T_Y_DISABLED}
\alias{module_emissions_L111.nonghg_en_R_S_T_Y_DISABLED}
\title{module_emissions_L111.nonghg_en_R_S_T_Y}
\usage{
module_emissions_L111.nonghg_en_R_S_T_Y_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L111.nonghg_tg_R_en_S_F_Yh}, \code{L111.nonghg_tgej_R_en_S_F_Yh}. The corresponding file in the
original data system was \code{L111.nonghg_en_R_S_T_Y.R} (emissions level1).
}
\description{
Calculate non-ghg emission totals and non-ghg emission shares of total emissions.
}
\details{
This code produces two outputs: non-ghg emission totals and non-ghg emission shares of total emissions.
First, non-ghg gas emissions are combined and grouped by sector and region, emissions are scaled, and international
shipping & aviation emission data calculated based on total emission and total emission shares. Finally, non-ghg emission
totals and shares are calculated by GCAM sector, fuel, technology, and driver type for EDGAR historical years.
}
\author{
CDL September 2017
}
