% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L112.ghg_en_R_S_T_Y.R
\name{module_emissions_L112.ghg_en_R_S_T_Y}
\alias{module_emissions_L112.ghg_en_R_S_T_Y}
\title{module_emissions_L112.ghg_en_R_S_T_Y}
\usage{
module_emissions_L112.ghg_en_R_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L112.ghg_tg_R_en_S_F_Yh}, \code{L112.ghg_tgej_R_en_S_F_Yh}. The corresponding file in the
original data system was \code{L112.ghg_en_R_S_T_Y.R} (emissions level1).
}
\description{
Calculates emissions and emissions factors using EPA emissions factors and scales to EDGAR emissions.
}
\details{
Calculates emissions using EPA emissions factors and energy data. Then scales to EDGAR emissions and calculates emissions factors.
}
\author{
RH July 2017
}
