% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L113.ghg_an_R_S_T_Y.R
\name{module_emissions_L113.ghg_an_R_S_T_Y}
\alias{module_emissions_L113.ghg_an_R_S_T_Y}
\title{module_emissions_L113.ghg_an_R_S_T_Y}
\usage{
module_emissions_L113.ghg_an_R_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L113.ghg_tg_R_an_C_Sys_Fd_Yh}. The corresponding file in the
original data system was \code{L113.ghg_an_R_S_T_Y.R} (emissions level1).
}
\description{
This chunk calculates the animal GHG emissions (CH4 and N2O)
 by GCAM region / sector / technology / historical year, by scaling EPA
 emissions by tech to match EDGAR
}
\details{
"Calculate the animal emissions (CH4 and N2O) first compute unscaled emissions
by country and technology, then match emissions factors from EPA, next
compute unscaled emissions (production * emfactors) and aggregate by sector and region,
then compute EDGAR emissions by region and sector and lastly, scale the EPA emissions
by tech to match EDGAR. Note that file L115, handles NH3."
}
\author{
CH July 2017
}
