% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L115.nh3_an_R_S_T_Y.R
\name{module_emissions_L115.nh3_an_R_S_T_Y}
\alias{module_emissions_L115.nh3_an_R_S_T_Y}
\title{module_emissions_L115.nh3_an_R_S_T_Y}
\usage{
module_emissions_L115.nh3_an_R_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L115.nh3_tg_R_an_C_Sys_Fd_Yh}. The corresponding file in the
original data system was \code{L115.nh3_an_R_S_T_Y.R} (emissions level1).
}
\description{
Annual animal NH3 emissions by GCAM region, sector and technology.
}
\details{
This chunk uses EPA emissions and FAO data to estimate agricultural NH3 emissions which are scaled to regional values using EDGAR data.
}
\author{
KD May 2017
}
