% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L1211.nonco2_awb_R_S_T_Y_IRR.R
\name{module_emissions_L1211.nonco2_awb_R_S_T_Y_IRR}
\alias{module_emissions_L1211.nonco2_awb_R_S_T_Y_IRR}
\title{module_emissions_L1211.nonco2_awb_R_S_T_Y_IRR}
\usage{
module_emissions_L1211.nonco2_awb_R_S_T_Y_IRR(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1211.nonco2_tg_R_awb_C_Y_GLU_IRR}, \code{L1211.ag_irrShare_R_C_Y_GLU_irr}. The corresponding file in the
original data system was \code{L1211.nonco2_awb_R_S_T_Y_IRR.R} (emissions level1).
}
\description{
This chunk produces irrigated and rainfed production shares, and agricultural waste burning emissions by irrigated & rainfed production shares.
}
\details{
First, we compute the irrigated and rainfed production shares within region, GLU, and crop.
Second, we sum the irrigated and rainfed production shares over the historical years to get the total by region, GLU, and crop.
Third, we divide the irrigated and rainfed production shares over the historical years to get the share of irrigated and rainfed production within region, GLU, and crop.
Finally, we multiply non-CO2 emissions within region, GLU, and crop by irrigated and rainfed production shares.
}
\author{
CDL April 2017
}
