% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L122.ghg_agr_R_S_T_Y.R
\name{module_emissions_L122.ghg_agr_R_S_T_Y}
\alias{module_emissions_L122.ghg_agr_R_S_T_Y}
\title{module_emissions_L122.ghg_agr_R_S_T_Y}
\usage{
module_emissions_L122.ghg_agr_R_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L122.EmissShare_R_C_Y_GLU}, \code{L122.ghg_tg_R_agr_C_Y_GLU}. The corresponding file in the
original data system was \code{L122.ghg_agr_R_S_T_Y} (emissions level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
