% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L1221.ghg_agr_R_S_T_Y_IRR.R
\name{module_emissions_L1221.ghg_agr_R_S_T_Y_IRR_DISABLED}
\alias{module_emissions_L1221.ghg_agr_R_S_T_Y_IRR_DISABLED}
\title{module_emissions_L1221.ghg_agr_R_S_T_Y_IRR}
\usage{
module_emissions_L1221.ghg_agr_R_S_T_Y_IRR_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1221.ghg_tg_R_agr_C_Y_GLU_IRR}. The corresponding file in the
original data system was \code{L1221.ghg_agr_R_S_T_Y_IRR.R} (emissions level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
