% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L124.nonco2_unmgd_R_S_T_Y.R
\name{module_emissions_L124.nonco2_unmgd_R_S_T_Y}
\alias{module_emissions_L124.nonco2_unmgd_R_S_T_Y}
\title{module_emissions_L124.nonco2_unmgd_R_S_T_Y}
\usage{
module_emissions_L124.nonco2_unmgd_R_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L124.nonco2_tg_R_grass_Y_GLU}, \code{L124.nonco2_tg_R_forest_Y_GLU}, \code{L124.deforest_coefs}. The corresponding file in the
original data system was \code{L124.nonco2_unmgd_R_S_T_Y.R} (emissions level1).
}
\description{
Calculate non-CO2 emissions from unmanaged lands (savanna burning, forest fires, deforestation)
}
\details{
Calculate non-CO2 emissions from unmanaged lands (savanna burning, forest fires, deforestation).
Downscale EDGAR regional emissions to GLU using shares of land area in each GLU within each region.
Divide forest-related emissions into deforestation and forest fires using GFED data
Compute global average deforestation emissions coefficients using deforestation from 2000 to 2005.
Note: File does not calculate emissions from BC/OC (separate chunk) or NH3 (seemingly omitted)
}
\author{
KVC May 2017
}
