% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L125.bcoc_unmgd_R_S_T_Y.R
\name{module_emissions_L125.bcoc_unmgd_R_S_T_Y_DISABLED}
\alias{module_emissions_L125.bcoc_unmgd_R_S_T_Y_DISABLED}
\title{module_emissions_L125.bcoc_unmgd_R_S_T_Y}
\usage{
module_emissions_L125.bcoc_unmgd_R_S_T_Y_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L125.bcoc_tgbkm2_R_grass_2000}, \code{L125.bcoc_tgbkm2_R_forest_2000}, \code{L125.deforest_coefs_bcoc}. The corresponding file in the
original data system was \code{L125.bcoc_unmgd_R_S_T_Y.R} (emissions level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
