% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L131.nonco2_proc_R_S_T_Y.R
\name{module_emissions_L131.nonco2_proc_R_S_T_Y}
\alias{module_emissions_L131.nonco2_proc_R_S_T_Y}
\title{module_emissions_L131.nonco2_proc_R_S_T_Y}
\usage{
module_emissions_L131.nonco2_proc_R_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L131.nonco2_tg_R_prc_S_S_Yh}. The corresponding file in the
original data system was \code{L131.nonco2_proc_R_S_T_Y.R} (emissions level1).
}
\description{
Calculate historical emissions from the processing sector by GCAM technology, computed from EDGAR emissions data and EPA emissions factors.
}
\details{
This chunk calculates the historical emissions from the processing sector by GCAM technology,
computed from EDGAR emissions data and EPA emissions factors. First: Compute subsector share of sectoral emissions using EPA data.
Second: Compute share of sectoral emissions in each subsector using EPA data. Third: Group by and then join by sector, subsector,
and Non.CO2. Fourth: Disaggregate EDGAR emissions to subsectors. Fifth: map in all data and compute emissions (EDGAR emissions*tech_share)
}
\author{
CH May 2017
}
