% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L151.ctrl_R_en_S_T.R
\name{module_emissions_L151.ctrl_R_en_S_T}
\alias{module_emissions_L151.ctrl_R_en_S_T}
\title{module_emissions_L151.ctrl_R_en_S_T}
\usage{
module_emissions_L151.ctrl_R_en_S_T(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L151.nonghg_ctrl_R_en_S_T}. The corresponding file in the
original data system was \code{L151.ctrl_R_en_S_T.R} (emissions level1).
}
\description{
Compute maximum reduction by region / sector / gas.
}
\details{
Compute maximum reduction by region and sector for SO2, CO, NOx, NMVOC (all 2005 reference), BC, and OC (2005).
Reductions are limited to [0, 100].
}
\author{
BBL April 2017
}
