% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L152.MACC.R
\name{module_emissions_L152.MACC_DISABLED}
\alias{module_emissions_L152.MACC_DISABLED}
\title{module_emissions_L152.MACC}
\usage{
module_emissions_L152.MACC_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L152.MAC_pct_R_S_Proc_EPA}. The corresponding file in the
original data system was \code{L152.MACC.R} (emissions level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
