% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L211.ag_nonco2.R
\name{module_emissions_L211.ag_nonco2}
\alias{module_emissions_L211.ag_nonco2}
\title{module_emissions_L211.ag_nonco2}
\usage{
module_emissions_L211.ag_nonco2(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L211.AWBEmissions}, \code{L211.AGREmissions}, \code{L211.AnEmissions}, \code{L211.AnNH3Emissions}, \code{L211.AGRBio}, \code{L211.AWB_BCOC_EmissCoeff}, \code{L211.nonghg_max_reduction}, \code{L211.nonghg_steepness}. The corresponding file in the
original data system was \code{L211.ag_nonco2.R} (emissions level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
RH July 2017
}
