% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2111.ag_nonco2_IRR.R
\name{module_emissions_L2111.ag_nonco2_IRR}
\alias{module_emissions_L2111.ag_nonco2_IRR}
\title{module_emissions_L2111.ag_nonco2_IRR}
\usage{
module_emissions_L2111.ag_nonco2_IRR(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2111.AWBEmissions}, \code{L2111.AGREmissions}, \code{L2111.AGRBio}, \code{L2111.AWB_BCOC_EmissCoeff}, \code{L2111.nonghg_max_reduction}, \code{L2111.nonghg_steepness}. The corresponding file in the
original data system was \code{L2111.ag_nonco2_IRR.R} (emissions level2).
}
\description{
Creates agricultural, agricultural waste burning and animal non-CO2 emissions.
Writes biomass N2O coefficients, ag waste burning BC/OC coefficients, non-GHG maximum coefficient reduction, and non-GHG steepness by agricultural technology.
}
\details{
Writes out agricultural, agricultural waste burning, and animal non-CO2 emissions. Writes biomass N2O coefficients,
ag waste burning BC/OC coefficients, non-GHG maximum coefficient reduction, and non-GHG steepness by agricultural technology.
}
\author{
RLH July 2017
}
