% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L241.en_newtech_nonco2.R
\name{module_emissions_L241.en_newtech_nonco2_DISABLED}
\alias{module_emissions_L241.en_newtech_nonco2_DISABLED}
\title{module_emissions_L241.en_newtech_nonco2}
\usage{
module_emissions_L241.en_newtech_nonco2_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L241.nonco2_tech_coeff}, \code{L241.nonghg_max_reduction}, \code{L241.nonghg_steepness}. The corresponding file in the
original data system was \code{L241.en_newtech_nonco2.R} (emissions level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
