% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L241.fgas.R
\name{module_emissions_L241.fgas_DISABLED}
\alias{module_emissions_L241.fgas_DISABLED}
\title{module_emissions_L241.fgas}
\usage{
module_emissions_L241.fgas_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L241.hfc_all}, \code{L241.pfc_all}, \code{L241.hfc_future}, \code{L241.fgas_all_units}. The corresponding file in the
original data system was \code{L241.fgas.R} (emissions level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
