% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L251.en_ssp_nonco2.R
\name{module_emissions_L251.en_ssp_nonco2}
\alias{module_emissions_L251.en_ssp_nonco2}
\title{module_emissions_L251.en_ssp_nonco2}
\usage{
module_emissions_L251.en_ssp_nonco2(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L251.ctrl.delete}, \code{L251.ssp15_ef}, \code{L251.ssp2_ef}, \code{L251.ssp34_ef}, \code{L251.ssp15_ef_vin}, \code{L251.ssp2_ef_vin}, \code{L251.ssp34_ef_vin}. The corresponding file in the
original data system was \code{L251.en_ssp_nonco2.R} (emissions level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
