% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2522.ag_MACC_IRR_MGMT.R
\name{module_emissions_L2522.ag_MACC_IRR_MGMT}
\alias{module_emissions_L2522.ag_MACC_IRR_MGMT}
\title{module_emissions_L2522.ag_MACC_IRR_MGMT --- driver is failing to merege, i think tha this has to do with a merge affecting another junk come back to it at a latter date.....}
\usage{
module_emissions_L2522.ag_MACC_IRR_MGMT(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2522.AgMAC}, \code{L2522.MAC_Ag_TC_SSP1}, \code{L2522.MAC_An_TC_SSP1}, \code{L2522.MAC_Ag_TC_SSP2}, \code{L2522.MAC_An_TC_SSP2}, \code{L2522.MAC_Ag_TC_SSP5}, \code{L2522.MAC_An_TC_SSP5}. The corresponding file in the
original data system was \code{L2522.ag_MACC_IRR_MGMT.R} (emissions level2).
}
\description{
Add new technology data from A_MACC_TechChange to the animal and agircultral marginal abatement cost "MAC" curves.
}
\details{
Add new technology data from A_MACC_TechChange to the animal and agircultral marginal abatement cost "MAC" curves.
}
\author{
KD Aug 2017
}
