% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_building_SSP.xml.R
\name{module_energy_batch_building_SSP.xml}
\alias{module_energy_batch_building_SSP.xml}
\title{module_energy_batch_building_SSP.xml}
\usage{
module_energy_batch_building_SSP.xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{building_SSP1.xml}, \code{building_SSP2.xml}, \code{building_SSP3.xml},
\code{building_SSP4.xml},and \code{building_SSP5.xml}.
}
\description{
Construct XML data structure for all the \code{building_SSP.xml} files.
}
