% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_industry_SSP_xml.R
\name{module_energy_batch_industry_SSP_xml}
\alias{module_energy_batch_industry_SSP_xml}
\title{module_energy_batch_industry_SSP_xml}
\usage{
module_energy_batch_industry_SSP_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{industry_GCAM3.xml}, \code{industry_SSP1.xml}, \code{industry_SSP2.xml}, \code{industry_SSP3.xml},
\code{industry_SSP4.xml}, \code{industry_SSP5.xml}, \code{industry_gSSP1.xml}, \code{industry_gSSP2.xml},
\code{industry_gSSP3.xml}, \code{industry_gSSP4.xml}, and \code{industry_gSSP5.xml}.
}
\description{
Construct XML data structures for all the \code{industry_SSP.xml} files.
}
