% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_resources_SSP_xml.R
\name{module_energy_batch_resources_SSP_xml_DISABLED}
\alias{module_energy_batch_resources_SSP_xml_DISABLED}
\title{module_energy_batch_resources_SSP_xml}
\usage{
module_energy_batch_resources_SSP_xml_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{resources_SSP5.xml}. The corresponding file in the
original data system was \code{batch_resources_SSP5_xml.R} (energy XML).
}
\description{
Construct XML data structure for \code{resources_SSP5.xml}.
}
