% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L202.Ccoef.R
\name{module_energy_L202.Ccoef_DISABLED}
\alias{module_energy_L202.Ccoef_DISABLED}
\title{module_energy_L202.Ccoef}
\usage{
module_energy_L202.Ccoef_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L202.CarbonCoef}. The corresponding file in the
original data system was \code{L202.Ccoef.R} (energy level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
