% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2321.cement.R
\name{module_energy_L2321.cement}
\alias{module_energy_L2321.cement}
\title{module_energy_L2321.cement}
\usage{
module_energy_L2321.cement(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2321.SectorLogitTables[[ curr_table ]]$data}, \code{L2321.Supplysector_cement}, \code{L2321.FinalEnergyKeyword_cement}, \code{L2321.SubsectorLogitTables[[ curr_table ]]$data}, \code{L2321.SubsectorLogit_cement}, \code{L2321.SubsectorShrwtFllt_cement}, \code{L2321.SubsectorInterp_cement}, \code{L2321.StubTech_cement}, \code{L2321.GlobalTechShrwt_cement}, \code{L2321.GlobalTechCoef_cement}, \code{L2321.GlobalTechCost_cement}, \code{L2321.GlobalTechCapture_cement}, \code{L2321.StubTechProd_cement}, \code{L2321.StubTechCalInput_cement_heat}, \code{L2321.StubTechCoef_cement}, \code{L2321.PerCapitaBased_cement}, \code{L2321.BaseService_cement}, \code{L2321.PriceElasticity_cement}, \code{object}. The corresponding file in the
original data system was \code{L2321.cement.R} (energy level2).
}
\description{
The chunk provides final energy keyword, supplysector/subsector information,
supplysector/subsector interpolation information, global technology share weight,
global technology efficiency, global technology coefficients, global technology cost,
price elasticity, stub technology information, stub technology interpolation information,
stub technology calibrated inputs, and etc for cement sector.
}
\details{
The chunk provides final energy keyword, supplysector/subsector information, supplysector/subsector interpolation information, global technology share weight, global technology efficiency, global technology coefficients, global technology cost, price elasticity, stub technology information, stub technology interpolation information, stub technology calibrated inputs, and etc for cement sector.
}
\author{
LF October 2017
}
