% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2322.Fert.R
\name{module_energy_L2322.Fert}
\alias{module_energy_L2322.Fert}
\title{module_energy_L2322.Fert}
\usage{
module_energy_L2322.Fert(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:  \code{L2322.Supplysector_Fert}, \code{L2322.FinalEnergyKeyword_Fert}, \code{L2322.SubsectorLogit_Fert}, \code{L2322.SubsectorShrwtFllt_Fert}, \code{L2322.SubsectorInterp_Fert}, \code{L2322.StubTech_Fert}, \code{L2322.GlobalTechShrwt_Fert}, \code{L2322.GlobalTechCoef_Fert}, \code{L2322.GlobalTechCost_Fert}, \code{L2322.GlobalTechCapture_Fert}, \code{L2322.GlobalTechSCurve_Fert}, \code{L2322.GlobalTechProfitShutdown_Fert}, \code{L2322.StubTechProd_Fert}, \code{L2322.StubTechCoef_Fert}, \code{L2322.StubTechFixOut_Fert_imp}, \code{L2322.StubTechFixOut_Fert_exp}, \code{L2322.PerCapitaBased_Fert}, \code{L2322.BaseService_Fert}. The corresponding file in the
original data system was \code{L2322.Fert.R} (energy level2).
}
\description{
Provide supply sector information/keywords, subsector shareweights, global technology lifetime,
energy inputs and coefficients, global fertilizer manufacturing technologies, etc. for the fertilizer sector.
}
\details{
This chunk provides supply sector information/keywords, subsector shareweights, global technology lifetime,
energy inputs and coefficients, global fertilizer manufacturing technologies, etc. for the fertilizer sector.
}
\author{
LF September 2017
}
