% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L242.building_agg.R
\name{module_energy_L242.building_agg_DISABLED}
\alias{module_energy_L242.building_agg_DISABLED}
\title{module_energy_L242.building_agg}
\usage{
module_energy_L242.building_agg_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L242.SectorLogitTables[[ curr_table ]]$data}, \code{L242.Supplysector_bld}, \code{L242.FinalEnergyKeyword_bld}, \code{L242.SubsectorLogitTables[[ curr_table ]]$data}, \code{L242.SubsectorLogit_bld}, \code{L242.SubsectorShrwt_bld}, \code{L242.SubsectorShrwtFllt_bld}, \code{L242.SubsectorInterp_bld}, \code{L242.SubsectorInterpTo_bld}, \code{L242.StubTech_bld}, \code{L242.GlobalTechInterp_bld}, \code{L242.GlobalTechShrwt_bld}, \code{L242.GlobalTechEff_bld}, \code{L242.GlobalTechCost_bld}, \code{L242.StubTechCalInput_bld}, \code{L242.FuelPrefElast_bld}, \code{L242.PerCapitaBased_bld}, \code{L242.PriceElasticity_bld}, \code{L242.BaseService_bld}. The corresponding file in the
original data system was \code{L242.building_agg.R} (energy level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
