% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L261.Cstorage.R
\name{module_energy_L261.Cstorage}
\alias{module_energy_L261.Cstorage}
\title{module_energy_L261.Cstorage}
\usage{
module_energy_L261.Cstorage(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L261.DepRsrc}, \code{L261.UnlimitRsrc}, \code{L261.DepRsrcCurves_C}, \code{L261.SectorLogitTables[[ curr_table ]]$data}, \code{L261.Supplysector_C}, \code{L261.SubsectorLogitTables[[ curr_table ]]$data}, \code{L261.SubsectorLogit_C}, \code{L261.SubsectorShrwtFllt_C}, \code{L261.StubTech_C}, \code{L261.GlobalTechCoef_C}, \code{L261.GlobalTechCost_C}, \code{L261.GlobalTechShrwt_C}, \code{L261.GlobalTechCost_C_High}, \code{L261.GlobalTechShrwt_C_nooffshore}, \code{L261.DepRsrcCurves_C_high}, \code{L261.DepRsrcCurves_C_low}, \code{L261.DepRsrcCurves_C_lowest}. The corresponding file in the
original data system was \code{L261.Cstorage.R} (energy level2).
}
\description{
Calculates carbon storage resource supply curves, shareweights, technology coefficients and costs, and other carbon storage information
}
\details{
The below fifteen tables pertaining to carbon storage properties are generated:
    Carbon storage information
    Unlimited carbon storage information
    Supply curve of carbon storage resources
    High supply curve of onshore carbon storage resources
    Low supply curve of onshore carbon storage resources
    Lowest supply curve of onshore carbon storage resources
    Carbon storage sector information
    Subsector logit exponents of carbon storage sector
    Subsector shareweights of carbon storage sectors
    Identification of stub technologies of carbon storage
    Carbon storage global technology coefficients across base model years
    Carbon storage global technology costs across base model years
    Carbon storage global technology costs across base model years, high price scenario
    Shareweights of carbon storage technologies across base model years
    Shareweights of offshore carbon storage technologies
}
\author{
AJS August 2017
}
