% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA100.CDIAC_downscale_ctry.R
\name{module_energy_LA100.CDIAC_downscale_ctry}
\alias{module_energy_LA100.CDIAC_downscale_ctry}
\title{module_energy_LA100.CDIAC_downscale_ctry}
\usage{
module_energy_LA100.CDIAC_downscale_ctry(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.CDIAC_CO2_ctry_hist}. The corresponding file in the
original data system was \code{LA100.CDIAC_downscale_ctry.R} (energy level1).
}
\description{
Combine and downscale (back in time, for USSR and Yugoslavia) the CDIAC emissions and sequestration data.
}
\details{
Combine and downscale (back in time, for USSR and Yugoslavia) the CDIAC emissions and sequestration data.
}
\author{
BBL April 2017
}
