% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA101.en_bal_IEA.R
\name{module_energy_LA101.en_bal_IEA}
\alias{module_energy_LA101.en_bal_IEA}
\title{module_energy_LA101.en_bal_IEA}
\usage{
module_energy_LA101.en_bal_IEA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L101.en_bal_EJ_R_Si_Fi_Yh_full}, \code{L101.en_bal_EJ_ctry_Si_Fi_Yh_full}, \code{L101.in_EJ_ctry_trn_Fi_Yh}, \code{L101.in_EJ_ctry_bld_Fi_Yh}. The corresponding file in the
original data system was \code{LA101.en_bal_IEA.R} (energy level1).
}
\description{
Rename IEA sectors and fuel data to nomenclature used in GCAM.
}
\details{
Rename IEA sectors and fuel data to nomenclature used in GCAM.
It also adjusts information uploaded from L100.IEA_en_bal_ctry_hist according to IEA_sector_fuel_modifications.
}
\author{
FF July 2017
}
