% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA101.en_bal_IEA.R
\name{module_energy_LA101.en_bal_IEA_DISABLED}
\alias{module_energy_LA101.en_bal_IEA_DISABLED}
\title{module_energy_LA101.en_bal_IEA}
\usage{
module_energy_LA101.en_bal_IEA_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L101.en_bal_EJ_R_Si_Fi_Yh_full}, \code{L101.en_bal_EJ_ctry_Si_Fi_Yh_full}, \code{L101.in_EJ_ctry_trn_Fi_Yh}, \code{L101.in_EJ_ctry_bld_Fi_Yh}. The corresponding file in the
original data system was \code{LA101.en_bal_IEA.R} (energy level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
