% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA1011.en_bal_adj.R
\name{module_energy_LA1011.en_bal_adj_DISABLED}
\alias{module_energy_LA1011.en_bal_adj_DISABLED}
\title{module_energy_LA1011.en_bal_adj}
\usage{
module_energy_LA1011.en_bal_adj_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1011.en_bal_EJ_R_Si_Fi_Yh}, \code{L1011.in_EJ_ctry_intlship_TOT_Yh}. The corresponding file in the
original data system was \code{LA1011.en_bal_adj.R} (energy level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
