% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA102.en_emiss_CDIAC.R
\name{module_energy_LA102.en_emiss_CDIAC}
\alias{module_energy_LA102.en_emiss_CDIAC}
\title{module_energy_LA102.en_emiss_CDIAC}
\usage{
module_energy_LA102.en_emiss_CDIAC(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L102.CO2_Mt_R_F_Yh}, \code{L102.Ccoef_kgCGJ_R_F_Yh}, \code{L102.Ccoef_kgCGJ_F_Yh}. The corresponding file in the
original data system was \code{LA102.en_emiss_CDIAC.R} (energy level1).
}
\description{
Calculates regional TPES as well as regional and global emissions coefficients by fuel
}
\details{
This chunk maps the CDIAC country emissions database (L100.CDIAC_CO2_ctry_hist) onto GCAM
regions (iso_GCAM_regID) and fuels (CDIAC_fuel) to determine regional emissions by fuel (L102.CO2_Mt_R_F_Yh),
extracts non-energy emissions and uses the remainder to calculate regional (L102.Ccoef_kgCGJ_R_F_Yh) and
global (L102.Ccoef_kgCGJ_F_Yh) emissions coefficients
}
\author{
CWR July 2017
}
