% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA102.en_emiss_CDIAC.R
\name{module_energy_LA102.en_emiss_CDIAC_DISABLED}
\alias{module_energy_LA102.en_emiss_CDIAC_DISABLED}
\title{module_energy_LA102.en_emiss_CDIAC}
\usage{
module_energy_LA102.en_emiss_CDIAC_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L102.CO2_Mt_R_F_Yh}, \code{L102.Ccoef_kgCGJ_R_F_Yh}, \code{L102.Ccoef_kgCGJ_F_Yh}. The corresponding file in the
original data system was \code{LA102.en_emiss_CDIAC.R} (energy level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
