% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA111.rsrc_fos_Prod.R
\name{module_energy_LA111.rsrc_fos_Prod}
\alias{module_energy_LA111.rsrc_fos_Prod}
\title{module_energy_LA111.rsrc_fos_Prod}
\usage{
module_energy_LA111.rsrc_fos_Prod(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L111.Prod_EJ_R_F_Yh}, \code{L111.RsrcCurves_EJ_R_Ffos}. The corresponding file in the
original data system was \code{LA111.rsrc_fos_Prod.R} (energy level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
