% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA113.MSW.R
\name{module_energy_LA113.MSW}
\alias{module_energy_LA113.MSW}
\title{module_energy_LA113.MSW}
\usage{
module_energy_LA113.MSW(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L113.RsrcCurves_EJ_R_MSW}. The corresponding file in the
original data system was \code{LA113.MSW.R} (energy level1).
}
\description{
Generate municipal solid waste data.
}
\details{
Describe in detail what this chunk does.
}
\author{
BBL
}
