% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA115.roofPV.R
\name{module_energy_LA115.roofPV}
\alias{module_energy_LA115.roofPV}
\title{module_energy_LA115.roofPV}
\usage{
module_energy_LA115.roofPV(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L115.RsrcCurves_EJ_R_roofPV}. The corresponding file in the
original data system was \code{LA115.roofPV.R} (energy level1).
}
\description{
This chunk converts rooftop PV resources from the 14 GCAM regions, i.e., "region_GCAM3," to the 32 GCAM region IDs.
}
\details{
Rooftop PV resources are given in the input file according to the 14 GCAM regions, i.e., "region_GCAM3." Because some regions span mutliple GCAM region IDs, population in 2010 was used to allocate proportionally.
}
\author{
AJS April 2017
}
