% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA116.geo.R
\name{module_energy_LA116.geo_DISABLED}
\alias{module_energy_LA116.geo_DISABLED}
\title{module_energy_LA116.geo}
\usage{
module_energy_LA116.geo_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L116.RsrcCurves_EJ_R_geo}, \code{L116.RsrcCurves_EJ_R_EGS}. The corresponding file in the
original data system was \code{LA116.geo.R} (energy level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
