% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA117.tradbio.R
\name{module_energy_LA117.tradbio}
\alias{module_energy_LA117.tradbio}
\title{module_energy_LA117.tradbio}
\usage{
module_energy_LA117.tradbio(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L117.RsrcCurves_EJ_R_tradbio}. The corresponding file in the
original data system was \code{LA117.tradbio.R} (energy level1).
}
\description{
Creates regional traditional biomass supply curves
}
\details{
Creates regional traditional biomass supply curves by multiplying the max historical amount
used in each region by the supply curve amount in the global assumption supply curve
}
\author{
RH March 2017
}
