% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA117.tradbio.R
\name{module_energy_LA117.tradbio_DISABLED}
\alias{module_energy_LA117.tradbio_DISABLED}
\title{module_energy_LA117.tradbio}
\usage{
module_energy_LA117.tradbio_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L117.RsrcCurves_EJ_R_tradbio}. The corresponding file in the
original data system was \code{LA117.tradbio.R} (energy level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
