% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA118.hydro.R
\name{module_energy_LA118.hydro}
\alias{module_energy_LA118.hydro}
\title{module_energy_LA118.hydro}
\usage{
module_energy_LA118.hydro(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L118.out_EJ_R_elec_hydro_Yfut}. The corresponding file in the
original data system was \code{LA118.hydro.R} (energy level1).
}
\description{
Calculate hydro potential in EJ from 2010 to 2100 by GCAM region ID
}
\details{
Different proxies are used to calculate hydro potential.

In most cases, a growth potential for each country was calculated, multiplied by its share in the region, and added to the base-year ouput
}
\author{
AS May 2017
}
