% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA121.oil.R
\name{module_energy_LA121.oil}
\alias{module_energy_LA121.oil}
\title{module_energy_LA121.oil}
\usage{
module_energy_LA121.oil(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L121.in_EJ_R_unoil_F_Yh}, \code{L121.in_EJ_R_TPES_crude_Yh}, \code{L121.in_EJ_R_TPES_unoil_Yh}. The corresponding file in the
original data system was \code{LA121.oil.R} (energy level1).
}
\description{
This chunk processes historical oil data and separates into unconventional oil, crude oil, and energy inputs to oil
}
\details{
This chunk uses energy production data to determine the energy inputs to the oil sector. It uses
the IEA energy balance to separate out unconventional oil production and crude oil (total liquids - unconventional oil).
}
\author{
JDH June 2017
}
