% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA123.electricity.R
\name{module_energy_LA123.electricity_DISABLED}
\alias{module_energy_LA123.electricity_DISABLED}
\title{module_energy_LA123.electricity}
\usage{
module_energy_LA123.electricity_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L123.out_EJ_R_elec_F_Yh}, \code{L123.in_EJ_R_elec_F_Yh}, \code{L123.eff_R_elec_F_Yh}, \code{L123.out_EJ_R_indchp_F_Yh}, \code{L123.in_EJ_R_indchp_F_Yh}. The corresponding file in the
original data system was \code{LA123.electricity.R} (energy level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
