% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA1231.elec_tech.R
\name{module_energy_LA1231.elec_tech_DISABLED}
\alias{module_energy_LA1231.elec_tech_DISABLED}
\title{module_energy_LA1231.elec_tech}
\usage{
module_energy_LA1231.elec_tech_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1231.in_EJ_R_elec_F_tech_Yh}, \code{L1231.out_EJ_R_elec_F_tech_Yh}, \code{L1231.eff_R_elec_F_tech_Yh}. The corresponding file in the
original data system was \code{LA1231.elec_tech.R} (energy level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
