% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA143.HDDCDD.R
\name{module_energy_LA143.HDDCDD_DISABLED}
\alias{module_energy_LA143.HDDCDD_DISABLED}
\title{module_energy_LA143.HDDCDD}
\usage{
module_energy_LA143.HDDCDD_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L143.HDDCDD_scen_R_Y}, \code{L143.HDDCDD_scen_RG3_Y}, \code{L143.HDDCDD_scen_ctry_Y}. The corresponding file in the
original data system was \code{LA143.HDDCDD.R} (energy level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
