% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA144.building_det_en.R
\name{module_energy_LA144.building_det_en_DISABLED}
\alias{module_energy_LA144.building_det_en_DISABLED}
\title{module_energy_LA144.building_det_en}
\usage{
module_energy_LA144.building_det_en_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L144.end_use_eff}, \code{L144.shell_eff_R_Y}, \code{L144.in_EJ_R_bld_serv_F_Yh}, \code{L144.NEcost_75USDGJ}, \code{L144.internal_gains}, \code{L144.base_service_EJ_serv}. The corresponding file in the
original data system was \code{LA144.building_det_en.R} (energy level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
