% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA144.building_det_flsp.R
\name{module_energy_LA144.building_det_flsp}
\alias{module_energy_LA144.building_det_flsp}
\title{module_energy_LA144.building_det_flsp}
\usage{
module_energy_LA144.building_det_flsp(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L144.flsp_bm2_R_res_Yh}, \code{L144.flsp_bm2_R_comm_Yh}, \code{L144.flspPrice_90USDm2_R_bld_Yh}. The corresponding file in the
original data system was \code{LA144.building_det_flsp.R} (energy level1).
}
\description{
This chunk calculates residential and commercial floorspace - and floorspace prices - by GCAM region and historical year.
}
\details{
Describe in detail what this chunk does.
}
\author{
AJS July 2017
}
