% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1322.Fert.R
\name{module_energy_LB1322.Fert}
\alias{module_energy_LB1322.Fert}
\title{module_energy_LB1322.Fert}
\usage{
module_energy_LB1322.Fert(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1322.Fert_Prod_MtN_R_F_Y}, \code{L1322.IO_R_Fert_F_Yh}, \code{L1322.in_EJ_R_indenergy_F_Yh}, \code{L1322.in_EJ_R_indfeed_F_Yh}, \code{L1322.Fert_NEcost_75USDkgN_F}. The corresponding file in the
original data system was \code{LB1322.Fert.R} (energy level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
AJS June 2017
}
