% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-gcam-usa.R
\name{module_gcam-usa_LA100.Socioeconomics}
\alias{module_gcam-usa_LA100.Socioeconomics}
\title{module_gcam-usa_LA100.Socioeconomics}
\usage{
"module_gcam-usa_LA100.Socioeconomics"(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}.
}
\description{
Construct the \code{gcam-usa} data structures.
}
\author{
BBL
}

