% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L223.electricity_USA.R
\name{module_gcam.usa_L223.electricity_USA_DISABLED}
\alias{module_gcam.usa_L223.electricity_USA_DISABLED}
\title{module_gcam.usa_L223.electricity_USA}
\usage{
module_gcam.usa_L223.electricity_USA_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L223.DeleteSubsector_USAelec}, \code{L223.SectorNodeEquiv}, \code{L223.TechNodeEquiv}, \code{L223.PassThroughSector_elec}, \code{L223.PassThroughTech_elec_FERC}, \code{object}, \code{L223.SectorLogitTables_USAelec[[ curr_table ]]$data}, \code{L223.Supplysector_USAelec}, \code{L223.SubsectorShrwtFllt_USAelec}, \code{L223.SubsectorInterp_USAelec}, \code{L223.SubsectorLogitTables_USAelec[[ curr_table ]]$data}, \code{L223.SubsectorLogit_USAelec}, \code{L223.TechShrwt_USAelec}, \code{L223.TechCoef_USAelec}, \code{L223.Production_USAelec}, \code{L223.SectorLogitTables_elec_FERC[[ curr_table ]]$data}, \code{L223.Supplysector_elec_FERC}, \code{L223.SubsectorShrwtFllt_elec_FERC}, \code{L223.SubsectorInterp_elec_FERC}, \code{L223.SubsectorLogitTables_elec_FERC[[ curr_table ]]$data}, \code{L223.SubsectorLogit_elec_FERC}, \code{L223.TechShrwt_elec_FERC}, \code{L223.TechCoef_elec_FERC}, \code{L223.Production_elec_FERC}, \code{L223.InterestRate_FERC}, \code{L223.Pop_FERC}, \code{L223.BaseGDP_FERC}, \code{L223.LaborForceFillout_FERC}, \code{L223.SubsectorShrwt_nuc_USA}, \code{L223.StubTechFixOut_elec_USA}, \code{L223.StubTechFixOut_hydro_USA}, \code{L223.StubTechProd_elec_USA}, \code{L223.StubTechMarket_elec_USA}, \code{L223.StubTechMarket_backup_USA}, \code{L223.StubTechElecMarket_backup_USA}, \code{L223.StubTechCapFactor_elec_wind_USA}, \code{L223.StubTechCapFactor_elec_solar_USA}. The corresponding file in the
original data system was \code{L223.electricity_USA.R} (gcam-usa level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
