% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2243.ren_soft_cal_USA.R
\name{module_gcam.usa_L2243.ren_soft_cal_USA}
\alias{module_gcam.usa_L2243.ren_soft_cal_USA}
\title{module_gcam.usa_L2243.ren_soft_cal_USA}
\usage{
module_gcam.usa_L2243.ren_soft_cal_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2243.StubTechFixOut_ren_soft_cal_2015_USA}, \code{L2243.GlobalTechShrwt_ren_soft_cal_2015_USA},
\code{L2243.PrimaryRenewKeyword_ren_soft_cal_2015_USA}.
The corresponding file in the original data system was \code{L2243.ren_soft_cal.R} (gcam-usa level2).
}
\description{
Create state-level fixed output for renewable (wind and solar) electricity generation in GCAM-USA with
multiple load segments to "soft calibrate" 2015 generation based on EIA historical data.
}
\details{
Create fixed outputs to "soft calibrate" 2015 electricity generation to historical data for GCAM-USA.
}
\author{
MTB September 2018
}
