% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L225.hydrogen_USA.R
\name{module_gcam.usa_L225.hydrogen_USA_DISABLED}
\alias{module_gcam.usa_L225.hydrogen_USA_DISABLED}
\title{module_gcam.usa_L225.hydrogen_USA}
\usage{
module_gcam.usa_L225.hydrogen_USA_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L225.DeleteSubsector_h2_USA}. The corresponding file in the
original data system was \code{L225.hydrogen_USA.R} (gcam-usa level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
