% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L232.industry_USA.R
\name{module_gcam.usa_L232.industry_USA}
\alias{module_gcam.usa_L232.industry_USA}
\title{module_gcam.usa_L232.industry_USA}
\usage{
module_gcam.usa_L232.industry_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L232.DeleteSupplysector_USAind}, \code{L232.DeleteFinalDemand_USAind}, \code{object}, \code{L232.StubTechCalInput_indenergy_USA}, \code{L232.StubTechCalInput_indfeed_USA}, \code{L232.StubTechProd_industry_USA}, \code{L232.StubTechCoef_industry_USA}, \code{L232.StubTechMarket_ind_USA}, \code{L232.StubTechSecMarket_ind_USA}, \code{L232.BaseService_ind_USA}. The corresponding file in the
original data system was \code{L232.industry_USA.R} (gcam-usa level2).
}
\description{
Prepare level 2 industry sector files for USA.
}
\details{
Prepare level 2 industry sector files for USA.
}
\author{
ST October 2017
}
