% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2322.Fert_USA.R
\name{module_gcam.usa_L2322.Fert_USA}
\alias{module_gcam.usa_L2322.Fert_USA}
\title{module_gcam.usa_L2322.Fert_USA}
\usage{
module_gcam.usa_L2322.Fert_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2322.DeleteSubsector_USAFert}, \code{L2322.FinalEnergyKeyword_USAFert}, \code{L2322.FinalEnergyKeyword_Fert_USA}, \code{L2322.StubTech_Fert_USA}, \code{L2322.SubsectorLogit_USAFert}, \code{L2322.SubsectorShrwtFllt_USAFert}, \code{L2322.TechShrwt_USAFert}, \code{L2322.Production_USAFert}, \code{L2322.TechCoef_USAFer}, \code{L2322.StubTechProd_Fert_USA}, \code{L2322.StubTechCoef_Fert_USA}, \code{L2322.StubTechMarket_Fert_USA}, \code{L2322.SubsectorLogit_Fert_USA}, \code{L2322.Supplysector_Fert_USA}, \code{L2322.SubsectorShrwtFllt_Fert_USA}, \code{L2322.SubsectorInterp_Fert_USA},\code{L2322.SubsectorInterp_USAFert}
 The corresponding file in the original data system was \code{L2322.Fert_USA.R} (gcam-usa level2).
}
\description{
Produce the logit tables for the fertilizer sector in GCAM-USA.
}
\details{
This chunk produces the logit tables for the fertilizer sector in GCAM-USA
}
\author{
KD October 2017
}
