% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L254.transportation_USA.R
\name{module_gcam.usa_L254.transportation_USA_DISABLED}
\alias{module_gcam.usa_L254.transportation_USA_DISABLED}
\title{module_gcam.usa_L254.transportation_USA}
\usage{
module_gcam.usa_L254.transportation_USA_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L254.DeleteSupplysector_USAtrn}, \code{L254.DeleteFinalDemand_USAtrn}, \code{object}, \code{L254.StubTranTechCalInput_USA}, \code{L254.StubTranTechProd_nonmotor_USA}, \code{L254.StubTranTechCalInput_passthru_USA}, \code{L254.BaseService_trn_USA}. The corresponding file in the
original data system was \code{L254.transportation_USA.R} (gcam-usa level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
