% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L270.limits_USA.R
\name{module_gcam.usa_L270.limits_USA}
\alias{module_gcam.usa_L270.limits_USA}
\title{module_gcam.usa_L270.limits_USA}
\usage{
module_gcam.usa_L270.limits_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L270.CreditMkt_USA},  \code{paste0( "L270.NegEmissBudget_USA_", c("GCAM3", paste0("SSP", 1:5), paste0("gSSP", 1:5), paste0("spa", 1:5)) )}.
}
\description{
Add the 50 states to the USA market in each of the L270 limits polices.  In
particular to limit the fraction of liquid feedstocks and inputs to electricity
generation which can come from sources other than crude oil.  Constrain the
total amount of subsidy as a fraction of GDP which an economy is will to give
to have net negative emissions.
}
\details{
Add 50 states to USA market for GCAM policy constraints which enforce limits
to liquid feedstocks and the amount of subsidies given for net negative emissions.
}
\author{
PLP June 2018
}
