% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA114.wind-gcamusa.R
\name{module_gcam.usa_LA114.Wind_DISABLED}
\alias{module_gcam.usa_LA114.Wind_DISABLED}
\title{module_gcam.usa_LA114.Wind}
\usage{
module_gcam.usa_LA114.Wind_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L114.CapacityFactor_wind_state}. The corresponding file in the
original data system was \code{LA114.Wind.R} (gcam-usa level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
