% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA115.RooftopPV.R
\name{module_gcam.usa_LA115.RooftopPV_DISABLED}
\alias{module_gcam.usa_LA115.RooftopPV_DISABLED}
\title{module_gcam.usa_LA115.RooftopPV}
\usage{
module_gcam.usa_LA115.RooftopPV_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L115.rsrc_state_rooftopPV}. The corresponding file in the
original data system was \code{LA115.RooftopPV.R} (gcam-usa level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
