% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA119.solar.R
\name{module_gcam.usa_LA119.Solar}
\alias{module_gcam.usa_LA119.Solar}
\title{module_gcam.usa_LA119.Solar}
\usage{
module_gcam.usa_LA119.Solar(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L119.CapFacScaler_PV_state}, \code{L119.CapFacScaler_CSP_state}. The corresponding file in the
original data system was \code{LA119.Solar.R} (gcam-usa level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
