% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA122.Refining.R
\name{module_gcam.usa_LA122.Refining}
\alias{module_gcam.usa_LA122.Refining}
\title{module_gcam.usa_LA122.Refining}
\usage{
module_gcam.usa_LA122.Refining(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L122.in_EJ_state_refining_F}, \code{L122.out_EJ_state_refining_F}. The corresponding file in the
original data system was \code{LA122.Refining.R} (gcam-usa level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
RLH September 2017
}
