% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA132.Industry-gcamusa.R
\name{module_gcam.usa_LA132.Industry}
\alias{module_gcam.usa_LA132.Industry}
\title{module_gcam.usa_LA132.Industry}
\usage{
module_gcam.usa_LA132.Industry(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L132.in_EJ_state_indnochp_F}, \code{L132.in_EJ_state_indchp_F}, \code{L132.out_EJ_state_indchp_F}, \code{L132.in_EJ_state_indfeed_F}. The corresponding file in the
original data system was \code{LA132.Industry.R} (gcam-usa level1).
}
\description{
Provides industrial energy consumption and industrial feedstock consumption by region/fuel/historical year.
}
\details{
Provides for each US state industrial energy consumption and industrial feedstock consumption by region/fuel/historical year.
}
\author{
ST September 2017
}
