% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA1322.Fert.R
\name{module_gcam.usa_LA1322.Fert_DISABLED}
\alias{module_gcam.usa_LA1322.Fert_DISABLED}
\title{module_gcam.usa_LA1322.Fert}
\usage{
module_gcam.usa_LA1322.Fert_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1322.out_Mt_state_Fert_Yh}, \code{L1322.IO_GJkg_state_Fert_F_Yh}, \code{L1322.in_EJ_state_Fert_Yh}. The corresponding file in the
original data system was \code{LA1322.Fert.R} (gcam-usa level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
