% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA142.Building.R
\name{module_gcam.usa_LA142.Building_DISABLED}
\alias{module_gcam.usa_LA142.Building_DISABLED}
\title{module_gcam.usa_LA142.Building}
\usage{
module_gcam.usa_LA142.Building_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L142.in_EJ_state_bld_F}. The corresponding file in the
original data system was \code{LA142.Building.R} (gcam-usa level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
