% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA143.HDDCDD-gcam-usa.R
\name{module_gcam.usa_LA143.HDDCDD_DISABLED}
\alias{module_gcam.usa_LA143.HDDCDD_DISABLED}
\title{module_gcam.usa_LA143.HDDCDD}
\usage{
module_gcam.usa_LA143.HDDCDD_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L143.share_state_Pop_CDD_sR9}, \code{L143.share_state_Pop_CDD_sR13}, \code{L143.share_state_Pop_HDD_sR9}, \code{L143.share_state_Pop_HDD_sR13}, \code{L143.HDDCDD_scen_state}. The corresponding file in the
original data system was \code{LA143.HDDCDD.R} (gcam-usa level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
