% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA144.Commercial.R
\name{module_gcam.usa_LA144.Commercial_DISABLED}
\alias{module_gcam.usa_LA144.Commercial_DISABLED}
\title{module_gcam.usa_LA144.Commercial}
\usage{
module_gcam.usa_LA144.Commercial_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L144.flsp_bm2_state_comm}, \code{L144.in_EJ_state_comm_F_U_Y}. The corresponding file in the
original data system was \code{LA144.Commercial.R} (gcam-usa level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
