% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA144.Residential.R
\name{module_gcam.usa_LA144.Residential}
\alias{module_gcam.usa_LA144.Residential}
\title{module_gcam.usa_LA144.Residential}
\usage{
module_gcam.usa_LA144.Residential(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L144.flsp_bm2_state_res}, \code{L144.in_EJ_state_res_F_U_Y}. The corresponding file in the
original data system was \code{LA144.Residential.R} (gcam-usa level1).
}
\description{
Calculates residential floorspace by state and residential energy consumption by state/fuel/end use.
}
\details{
Calculates residential floorspace by state and residential energy consumption by state/fuel/end use.
}
\author{
RLH September 2017
}
