% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA161.Cstorage-gcamusa.R,
%   R/zchunk_LA161.Cstorage.R
\name{module_gcam.usa_LA161.Cstorage_DISABLED}
\alias{module_gcam.usa_LA161.Cstorage_DISABLED}
\alias{module_gcam.usa_LA161.Cstorage_DISABLED}
\title{module_gcam.usa_LA161.Cstorage}
\usage{
module_gcam.usa_LA161.Cstorage_DISABLED(command, ...)

module_gcam.usa_LA161.Cstorage_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}

\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L161.Cstorage_FERC}. The corresponding file in the
original data system was \code{LA161.Cstorage.R} (gcam-usa level1).

Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L161.Cstorage_FERC}. The corresponding file in the
original data system was \code{LA161.Cstorage.R} (gcam-usa level1).
}
\description{
Briefly describe what this chunk does.

Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.

Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017

YourInitials CurrentMonthName 2017
}
