% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA171.nonco2_trn_USA_S_T_Y.R
\name{module_gcam.usa_LA171.nonco2_trn_USA_S_T_Y.R}
\alias{module_gcam.usa_LA171.nonco2_trn_USA_S_T_Y.R}
\title{module_gcam.usa_LA171.nonco2_trn_USA_S_T_Y.R}
\usage{
module_gcam.usa_LA171.nonco2_trn_USA_S_T_Y.R(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{}, \code{},
original data system was \code{LA171.nonco2_trn_USA_S_T_Y.R} (gcam-usa level2).
}
\description{
Generates GCAM-USA model inputs for transportation sector by states.
}
\details{
Generates GCAM-USA model inputs for transportation sector by states.
}
\author{
KD Aug 2018
}
