% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB123.Electricity.R
\name{module_gcam.usa_LB123.Electricity_DISABLED}
\alias{module_gcam.usa_LB123.Electricity_DISABLED}
\title{module_gcam.usa_LB123.Electricity}
\usage{
module_gcam.usa_LB123.Electricity_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L123.in_EJ_state_elec_F}, \code{L123.out_EJ_state_elec_F}, \code{L123.in_EJ_state_ownuse_elec}, \code{L123.out_EJ_state_ownuse_elec}. The corresponding file in the
original data system was \code{LB123.Electricity.R} (gcam-usa level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
