% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1231.Elec_tech.R
\name{module_gcam.usa_LB1231.Elec_tech}
\alias{module_gcam.usa_LB1231.Elec_tech}
\title{module_gcam.usa_LB1231.Elec_tech}
\usage{
module_gcam.usa_LB1231.Elec_tech(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1231.in_EJ_state_elec_F_tech}, \code{L1231.out_EJ_state_elec_F_tech}. The corresponding file in the
original data system was \code{LB1231.Elec_tech.R} (gcam-usa level1).
}
\description{
Downscaling electricity by state/fuel to state/fuel/technology
}
\details{
Downscaling electricity by state/fuel to state/fuel/technology
}
\author{
RLH August 2017
}
