% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1233.Elec_water.R
\name{module_gcam.usa_LB1233.Elec_water_DISABLED}
\alias{module_gcam.usa_LB1233.Elec_water_DISABLED}
\title{module_gcam.usa_LB1233.Elec_water}
\usage{
module_gcam.usa_LB1233.Elec_water_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1233.out_EJ_state_elec_F_tech_cool}, \code{L1233.share_sR_elec_F_tech_cool}, \code{L1233.wdraw_km3_state_elec}, \code{L1233.wcons_km3_state_elec}. The corresponding file in the
original data system was \code{LB1233.Elec_water.R} (gcam-usa level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
