% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zchunk_batch_negative_emissions_budget_usa_xml.R
\name{module_gcamusa_batch_negative_emissions_budget_usa_xml}
\alias{module_gcamusa_batch_negative_emissions_budget_usa_xml}
\title{module_gcamusa_batch_negative_emissions_budget_usa_xml}
\usage{
module_gcamusa_batch_negative_emissions_budget_usa_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{paste0("negative_emissions_budget_USA_", c("GCAM3", paste0("SSP", 1:5), paste0("gSSP", 1:5), paste0("spa", 1:5)), ".xml")}.
}
\description{
Construct XML data structure for \code{paste0("negative_emissions_budget_USA_", c("GCAM3", paste0("SSP", 1:5), paste0("gSSP", 1:5)), ".xml")}.
}
