% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L103.water_mapping_USA.R
\name{module_gcamusa_L103.water_mapping_USA}
\alias{module_gcamusa_L103.water_mapping_USA}
\title{module_gcamusa_L103.water_mapping_USA}
\usage{
module_gcamusa_L103.water_mapping_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L103.water_mapping_R_GLU_B_W_Ws_share},
\code{L103.water_mapping_R_B_W_Ws_share}. There was no corresponding file in the
original data system.
}
\description{
Calculate percentage shares to map water demands from USA regional level to state and basin.
}
\details{
Water demands by USA region / sector to basin and state.
}
\author{
NTG Oct 2019
}
