% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L203.water_mapping_USA.R
\name{module_gcamusa_L203.water_mapping_USA}
\alias{module_gcamusa_L203.water_mapping_USA}
\title{module_gcamusa_L203.water_mapping_USA}
\usage{
module_gcamusa_L203.water_mapping_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\description{
Mapping of water consumption/withdrawal to sectoral demands at the state level.
}
\details{
Generates water mapping sector input files to group demands by sectors.
}
\author{
NTG May 2020
}
