% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2231.nonewcoal_USA.R
\name{module_gcamusa_L2231.nonewcoal_USA}
\alias{module_gcamusa_L2231.nonewcoal_USA}
\title{module_gcamusa_L2231.nonewcoal_USA}
\usage{
module_gcamusa_L2231.nonewcoal_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2231.StubTechShrwt_nonewcoal_USA}, \code{L2231.StubTechShrwt_coal_delay_USA}.
The corresponding file in the
original data system was \code{L2231.nonewcoal_USA.R} (gcam-usa level2).
}
\description{
Generates optional moratorium on new pulverized coal plants in USA states.
}
\details{
This chunk sets zero share-weights of pulverized coal technologies, which assumes
no new pulverized coal plants without CCS will be built in USA states.
}
\author{
RC Aug 2018
}
