% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2244.nuclear_USA.R
\name{module_gcamusa_L2244.nuclear_USA}
\alias{module_gcamusa_L2244.nuclear_USA}
\title{module_gcamusa_L2244.nuclear_USA}
\usage{
module_gcamusa_L2244.nuclear_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2244.StubTechSCurve_nuc_gen2_USA}.
The corresponding file in the original data system was \code{L2244.nuclear_USA.R} (gcam-usa level2).
}
\description{
Generates an add-on file to update nuclear assumptions in GCAM-USA.
}
\details{
This chunk creates an add-on file to update nuclear assumptions in GCAM-USA. Specifically, it reads in state-specific
s-curve retirement functions and lifetimes for existing nuclear vintage based on data put together by Dr. Son H. Kim.
The data reflect planned retirements by state and nuclear power plant. Further, the script updates share-weight assumptions for Gen 3 technology.
It assumes that no new nuclear is deployed in any state through 2035, except in GA, based on current understanding. It does not update
subsector share-weight assumptions in this add-on.
}
\author{
RC Sep 2018; edited MB Sep 2018
}
