% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA115.rooftopPV_USA.R
\name{module_gcamusa_LA115.rooftopPV}
\alias{module_gcamusa_LA115.rooftopPV}
\title{module_gcam.usa_LA115.RooftopPV}
\usage{
module_gcamusa_LA115.rooftopPV(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L115.rsrc_state_rooftopPV}. The corresponding file in the
original data system was \code{LA115.RooftopPV.R} (gcam-usa level1).
}
\description{
Prepare resource curves for rooftop PV (commercial and residential combined).
}
\details{
Prepare resource curves for rooftop PV (commercial and residential combined).
}
\author{
ST September 2017
}
