% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA119.solar_USA.R
\name{module_gcamusa_LA119.solar}
\alias{module_gcamusa_LA119.solar}
\title{module_gcam.usa_LA119.Solar}
\usage{
module_gcamusa_LA119.solar(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L119.CapFacScaler_PV_state}, \code{L119.CapFacScaler_CSP_state}. The corresponding file in the
original data system was \code{LA119.Solar.R} (gcam-usa level1).
}
\description{
Compute scalars by state to vary capacity factors by state.
}
\details{
This chunk computes scalars by state to vary capacity factors for central station PV and CSP technologies by state.
}
\author{
GI, FF, AS Apr 2017
}
