% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA154.Transport.R
\name{module_gcamusa_LA154.Transport}
\alias{module_gcamusa_LA154.Transport}
\title{module_gcam.usa_LA154.Transport}
\usage{
module_gcamusa_LA154.Transport(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L154.in_EJ_state_trn_m_sz_tech_F}, \code{L154.out_mpkm_state_trn_nonmotor_Yh}, \code{L154.in_EJ_state_trn_F}. The corresponding file in the
original data system was \code{LA154.Transport.R} (gcam-usa level1).
}
\description{
Downscale transportation energy consumption and nonmotor data to the state level, generating three ouput tables.
}
\details{
Transportation energy data was downscaled in proportion to EIA state-level transportation energy data

Transportation nonmotor data was downscaled in proportion to state population
}
\author{
AJS June 2017
}
