% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA2233.electricity_water_USA_test.R
\name{module_gcamusa_LA2233.electricity_water_USA_test}
\alias{module_gcamusa_LA2233.electricity_water_USA_test}
\title{module_gcamusa_LA2233.electricity_water_USA_test}
\usage{
module_gcamusa_LA2233.electricity_water_USA_test(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2233.StubTech_WaterCoef_ref}
The corresponding file in the
original data system was \code{LA2233.electricity_water_USA} (gcam-usa level2)
}
\description{
Weighted water coefficient for reference scenario and load segment classification
}
\details{
Weighted water coefficient for reference scenario and load segment classification
}
\author{
Zarrar Khan September 2018, NTG Aug 2019
}
