% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_modeltime.xml.R
\name{module_modeltime_batch_modeltime.xml}
\alias{module_modeltime_batch_modeltime.xml}
\title{module_modeltime_batch_modeltime.xml}
\usage{
module_modeltime_batch_modeltime.xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{modeltime.xml}. The corresponding file in the
original data system was \code{batch_modeltime.xml.R} (modeltime XML).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
