% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_modeltime_xml.R
\name{module_modeltime_batch_modeltime_xml}
\alias{module_modeltime_batch_modeltime_xml}
\title{module_modeltime_batch_modeltime_xml}
\usage{
module_modeltime_batch_modeltime_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{modeltime.xml}. The corresponding file in the
original data system was \code{batch_modeltime_xml.R} (modeltime XML).
}
\description{
Construct XML data structure for \code{modeltime.xml}.
}
