% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime.R
\name{module_modeltime_L200.modeltime}
\alias{module_modeltime_L200.modeltime}
\title{module_modeltime_L200.modeltime}
\usage{
module_modeltime_L200.modeltime(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}
}
\description{
Construct the \code{modeltime} data structures
}
\examples{
module_modeltime("MAKE", all_data = NULL)
}
\author{
BBL
}

