% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-chunk.R
\name{module_sample_sample}
\alias{module_sample_sample}
\title{module_sample_sample}
\usage{
module_sample_sample(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}.
}
\description{
Briefly describe what this chunk does.
}
\author{
Author name
}

