% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_interest_rate_xml.R
\name{module_socio_batch_interest_rate_xml}
\alias{module_socio_batch_interest_rate_xml}
\title{module_socio_batch_interest_rate_xml}
\usage{
module_socio_batch_interest_rate_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{interest_rate.xml}. The corresponding file in the
original data system was \code{batch_interest_rate.xml.R} (socio XML).
}
\description{
Construct XML data structure for \code{interest_rate.xml}.
}
