% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_socioeconomics_SSP_xml.R
\name{module_socio_batch_SSP.xml}
\alias{module_socio_batch_SSP.xml}
\title{module_socio_batch_SSP.xml}
\usage{
module_socio_batch_SSP.xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{socioeconomics_gSSP1.xml}, \code{socioeconomics_gSSP2.xml}, \code{socioeconomics_gSSP3.xml},
\code{socioeconomics_gSSP4.xml}, \code{socioeconomics_gSSP5.xml}, \code{socioeconomics_SSP1.xml},
\code{socioeconomics_SSP2.xml}, \code{socioeconomics_SSP3.xml},
\code{socioeconomics_SSP4.xml}, and \code{socioeconomics_SSP5.xml}.
}
\description{
Construct XML data structure for all the \code{socioeconomics_[g]SSP[1-5].xml} files.
}
