% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_socioeconomics_SSP2_xml.R
\name{module_socio_batch_SSP2.xml}
\alias{module_socio_batch_SSP2.xml}
\title{module_socio_batch_SSP2.xml}
\usage{
module_socio_batch_SSP2.xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{socioeconomics_SSP2.xml}. The corresponding file in the
original data system was \code{batch_socioeconomics_SSP2.xml} (socio XML).
}
\description{
Construct XML data structure for \code{socioeconomics_SSP2.xml}.
}
