% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeconomics.R
\name{module_socioeconomics_inputs}
\alias{module_socioeconomics_inputs}
\title{module_socioeconomics_inputs}
\usage{
module_socioeconomics_inputs(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}
}
\description{
Load the \code{socioeconomics} input data
}
\author{
BBL
}

