% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeconomics.R
\name{module_socioeconomics_L100.GDP_hist}
\alias{module_socioeconomics_L100.GDP_hist}
\title{module_socioeconomics_L100.GDP_hist}
\usage{
module_socioeconomics_L100.GDP_hist(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}.
}
\description{
Construct the \code{socioeconomics} data structures.
}
\author{
BBL
}

