% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-socioeconomics.R
\name{module_socioeconomics_L100.Population_downscale_ctry_DISABLED}
\alias{module_socioeconomics_L100.Population_downscale_ctry_DISABLED}
\title{module_socioeconomics_L100.Population_downscale_ctry_DISABLED}
\usage{
module_socioeconomics_L100.Population_downscale_ctry_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}.
}
\description{
Construct the \code{L100.Population_downscale_ctry} data structures.
}
\author{
BBL
}

