% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-socioeconomics.R
\name{module_socioeconomics_L101.Population}
\alias{module_socioeconomics_L101.Population}
\title{module_socioeconomics_L101.Population}
\usage{
module_socioeconomics_L101.Population(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}.
}
\description{
Construct the \code{socioeconomics} data structures.
}
\author{
BBL
}

