% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L242.Bld_Inc_Elas_scenarios.R
\name{module_socioeconomics_L242.Bld_Inc_Elas_scenarios_DISABLED}
\alias{module_socioeconomics_L242.Bld_Inc_Elas_scenarios_DISABLED}
\title{module_socioeconomics_L242.Bld_Inc_Elas_scenarios}
\usage{
module_socioeconomics_L242.Bld_Inc_Elas_scenarios_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L242.IncomeElasticity_bld_GCAM3}, \code{object}. The corresponding file in the
original data system was \code{L242.Bld_Inc_Elas_scenarios.R} (socioeconomics level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
