% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_MACC_TC_SSP1_IRR_MGMT.xml.R,
%   R/zchunk_batch_MACC_TC_SSP2_IRR_MGMT.xml.R,
%   R/zchunk_batch_MACC_TC_SSP5_IRR_MGMT.xml.R
\name{module_water_batch_water_mapping.xml_DISABLED}
\alias{module_water_batch_water_mapping.xml_DISABLED}
\alias{module_water_batch_water_mapping.xml_DISABLED}
\alias{module_water_batch_water_mapping.xml_DISABLED}
\title{module_water_batch_water_mapping.xml}
\usage{
module_water_batch_water_mapping.xml_DISABLED(command, ...)

module_water_batch_water_mapping.xml_DISABLED(command, ...)

module_water_batch_water_mapping.xml_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}

\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}

\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{water_mapping.xml}. The corresponding file in the
original data system was \code{batch_water_mapping.xml.R} (water XML).

Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{water_mapping.xml}. The corresponding file in the
original data system was \code{batch_water_mapping.xml.R} (water XML).

Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{water_mapping.xml}. The corresponding file in the
original data system was \code{batch_water_mapping.xml.R} (water XML).
}
\description{
Construct XML data structure for \code{water_mapping.xml}.

Construct XML data structure for \code{water_mapping.xml}.

Construct XML data structure for \code{water_mapping.xml}.
}
