% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L103.water.basin_mapping.R
\name{module_water_L103.water.basin_mapping}
\alias{module_water_L103.water.basin_mapping}
\title{module_water_L103.water.basin_mapping}
\usage{
module_water_L103.water.basin_mapping(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L103.water_mapping_R_GLU_B_W_Ws_share},
\code{L103.water_mapping_R_B_W_Ws_share}. There was no corresponding file in the
original data system.
}
\description{
Calculate percentage shares to map water demands by region / sector to basin.
}
\details{
Water demands to by region / sector to basin.
}
\author{
ST Oct 2018
}
