% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L110.water.demand.primary.R
\name{module_water_L110.water.demand.primary}
\alias{module_water_L110.water.demand.primary}
\title{module_water_L110.water.demand.primary}
\usage{
module_water_L110.water.demand.primary(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L110.water_demand_primary_R_S_W_m3_GJ}. The corresponding file in the
original data system was \code{L110.water.demand.primary.R} (water level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
