% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L133.water.demand.livestock.R
\name{module_water_L133.water.demand.livestock}
\alias{module_water_L133.water.demand.livestock}
\title{module_water_L133.water.demand.livestock}
\usage{
module_water_L133.water.demand.livestock(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L133.water_demand_livestock_R_C_W_km3_Mt}. The corresponding file in the
original data system was \code{L133.water.demand.livestock.R} (water level1).
}
\description{
Calculate livestock water coefficients by region ID / GCAM_commodity/ water type
}
\details{
Describe in detail what this chunk does.
}
\author{
KRD November 2017
}
