% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L202.water.resources.unlimited.R
\name{module_water_L202.water.resources.unlimited}
\alias{module_water_L202.water.resources.unlimited}
\title{module_water_L202.water.resources.unlimited}
\usage{
module_water_L202.water.resources.unlimited(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L202.UnlimitRsrc}, \code{L202.UnlimitRsrcPrice}. The corresponding file in the
original data system was \code{L202.water.resources.unlimited.R} (water level2).
}
\description{
Create unlimited resource markets for water types, and read in fixed prices for water types.
}
\details{
Create unlimited resource markets (i.e., 32 GCAM regions) for water types (i.e., water consumption, withdrawals, biophysical water consumption and seawater),
and read in fixed prices for water types.
}
\author{
YL July 2017
}
