% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L203.water.mapping.R
\name{module_water_L203.water.mapping}
\alias{module_water_L203.water.mapping}
\title{module_water_L203.water.mapping}
\usage{
module_water_L203.water.mapping(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L203.Supplysector}, \code{L203.SubsectorLogit}, \code{L203.SubsectorShrwtFllt}, \code{L203.TechShrwt}, \code{L203.TechCoef},
\code{L203.TechPmult}, \code{L203.TechDesalCoef}, \code{L203.TechDesalShrwt}, \code{L203.TechDesalCost}. The corresponding file in the
original data system was \code{L203.water.mapping.R} (water level2).
}
\description{
Mapping of water consumption/withdrawal to sectoral demands.
}
\details{
Generates water mapping sector input files to group demands by sectors.
}
\author{
ST August 2017 / ST Oct 2018
}
