% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L203.water.mapping.R
\name{module_water_L203.water.mapping_DISABLED}
\alias{module_water_L203.water.mapping_DISABLED}
\title{module_water_L203.water.mapping}
\usage{
module_water_L203.water.mapping_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L203.SectorLogitTables[[ curr_table ]]$data}, \code{L203.Supplysector}, \code{L203.SubsectorLogitTables[[ curr_table ]]$data}, \code{L203.SubsectorLogit}, \code{L203.SubsectorShrwtFllt}, \code{L203.TechShrwt}, \code{L203.TechCoef}. The corresponding file in the
original data system was \code{L203.water.mapping.R} (water level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
