% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L232.water.demand.manufacturing.R
\name{module_water_L232.water.demand.manufacturing_DISABLED}
\alias{module_water_L232.water.demand.manufacturing_DISABLED}
\title{module_water_L232.water.demand.manufacturing}
\usage{
module_water_L232.water.demand.manufacturing_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L232.TechCoef}. The corresponding file in the
original data system was \code{L232.water.demand.manufacturing.R} (water level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
