% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L232.water_demand_manufacturing.R
\name{module_water_L232.water_demand_manufacturing}
\alias{module_water_L232.water_demand_manufacturing}
\title{module_water_L232.water_demand_manufacturing}
\usage{
module_water_L232.water_demand_manufacturing(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L232.TechCoef}. The corresponding file in the
original data system was \code{L232.water_demand_manufacturing.R} (water level2).
}
\description{
Computes manufacturing water withdrawal/consumption coefficients (m3/GJ output) by region and year
}
\details{
Water widthdrawal and consumption coefficients for manufacturing for GCAM regions.
}
\author{
GPK June 2018
}
