% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L233.water.demand.livestock.R
\name{module_water_L233.water.demand.livestock_DISABLED}
\alias{module_water_L233.water.demand.livestock_DISABLED}
\title{module_water_L233.water.demand.livestock}
\usage{
module_water_L233.water.demand.livestock_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L233.TechCoef}. The corresponding file in the
original data system was \code{L233.water.demand.livestock.R} (water level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
