% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L245.water.demand.municipal.R
\name{module_water_L245.water.demand.municipal_DISABLED}
\alias{module_water_L245.water.demand.municipal_DISABLED}
\title{module_water_L245.water.demand.municipal}
\usage{
module_water_L245.water.demand.municipal_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L245.SectorLogitTables[[ curr_table ]]$data}, \code{L245.Supplysector}, \code{L245.SubsectorLogitTables[[ curr_table ]]$data}, \code{L245.SubsectorLogit}, \code{L245.SubsectorShrwtFllt}, \code{L245.TechShrwt}, \code{L245.TechCoef}, \code{L245.TechCost}, \code{L245.PerCapitaBased}, \code{L245.BaseService}, \code{L245.IncomeElasticity}, \code{L245.PriceElasticity}, \code{L245.aeei}. The corresponding file in the
original data system was \code{L245.water.demand.municipal.R} (water level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
