% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L245.water_demand_municipal.R
\name{module_water_L245.water_demand_municipal}
\alias{module_water_L245.water_demand_municipal}
\title{module_water_L245.water_demand_municipal}
\usage{
module_water_L245.water_demand_municipal(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L245.Supplysector}, \code{L245.SubsectorLogit}, \code{L245.SubsectorShrwtFllt}, \code{L245.TechShrwt}, \code{L245.TechCoef}, \code{L245.TechCost}, \code{L245.PerCapitaBased}, \code{L245.BaseService}, \code{L245.IncomeElasticity}, \code{L245.PriceElasticity}, \code{L245.aeei}. The corresponding file in the
original data system was \code{L245.water_demand_municipal.R} (water level2).
}
\description{
Expands municipal water information (cost, efficiency, coefficients) across regions and model years
}
\details{
Describe in detail what this chunk does.
}
\author{
ST August 2017
}
