% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_files}
\alias{normalize_files}
\title{normalize_files}
\usage{
normalize_files(root = system.file("extdata", package = "gcamdata"),
  min_compress_size = 1)
}
\arguments{
\item{root}{Folder root to scan, character}

\item{min_compress_size}{Files above this size (numeric, in MB) will be compressed when re-written}
}
\value{
Nothing - run for side effects only.
}
\description{
Normalize line endings for all package input data and compress files.
}
\details{
Some GCAM input datafiles have bad line endings (like you would find in Mac OS 9 and previous), and/or
don't have a final newline. This utility script converts all files to have Unix line endings (\code{\\n}) and a final newline.
It also automatically compresses files above a certain size.
}
\note{
Set \code{root} to "./extdata" in the git directory, not the package root, to make changes that 'stick'.
}
\author{
BBL
}
