% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outputs_of}
\alias{outputs_of}
\title{outputs_of}
\usage{
outputs_of(chunks)
}
\arguments{
\item{chunks}{Names of chunks, character}
}
\value{
Character vector of inputs.
}
\description{
Convenience function for getting the outputs of one or more chunks
}
