% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_csv_header}
\alias{parse_csv_header}
\title{parse_csv_header}
\usage{
parse_csv_header(obj, filename, n = 20, enforce_requirements = FALSE)
}
\arguments{
\item{obj}{The object to attach attributes to}

\item{filename}{Fully-qualified filename}

\item{n}{Number of lines to read from the beginning of the file}

\item{enforce_requirements}{Enforce mandatory fields?}
}
\value{
An empty \code{\link{tibble}} with appropriate attributes filled in.
}
\description{
Parse a CSV file's header, if present.
}
\details{
Headers are given at the top of files and consist of labels ("Title:", "Units:", etc)
prefixed by comment characters (#). The parser looks for these, and calls \code{\link{add_title}} and
similar functions to return an empty data frame with appropriate attribute set.
}
