% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{prebuilt_data}
\alias{prebuilt_data}
\title{prebuilt_data}
\usage{
prebuilt_data(object_name)
}
\arguments{
\item{object_name}{The name of the desired object, character}
}
\value{
The data object (a tibble).
}
\description{
Extract a prebuilt data object from the PREBUILT_DATA store.
}
