% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{protect_float}
\alias{protect_float}
\title{Protect floating point values in a data frame.}
\usage{
protect_float(df)
}
\arguments{
\item{df}{Data frame to have floats protected.}
}
\value{
Data frame with floating point columns converted to character.
}
\description{
All of the currently extant functions for writing tables render floating
point values using C's %g format.  This format is badly broken, in that a
value like 2.0 will be output as "2".  This can cause problems when the table
is read in at some later time, as "2" looks like an integer.  This function
converts all floating point columns in a data frame to character strings that
are properly formatted as floating point literals.  Using
\code{\link[readr]{write_csv}} on a data frame protected in this way will
produce the expected output.
}
\note{
The output produced this way will probably be a lot larger than one
produced by the default behavior, so this function should be used only where
the default behavior is causing problems.

Because the numeric columns are converted to characters, a data
frame converted in this way is no longer useful for computation.
}
