% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{read_logit_fn_tables}
\alias{read_logit_fn_tables}
\title{read_logit_fn_tables}
\usage{
read_logit_fn_tables(domain, table.name, skip = 0,
  include.equiv.table = FALSE)
}
\arguments{
\item{domain}{The data domain to read the table from}

\item{table.name}{The base table name to generate all logit type table names}

\item{skip}{Logical: whether to skip any lines (such as head information) when reading the table}

\item{include.equiv.table}{Logical: whether to include the EQUIV_TABLE in the list of results}
}
\value{
The tables TODO NEEED BETTER DOCUMENTATION
}
\description{
Read all logit function type tables
}
\details{
Read all logit function type tables given a base table name and return them in a list
where the full table name -> the read data.
}
