% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{remove_data}
\alias{remove_data}
\title{remove_data}
\usage{
remove_data(data_list, all_data)
}
\arguments{
\item{data_list}{A character vector of object names}

\item{all_data}{An existing (possibly empty) data store}
}
\value{
The modified data store.
}
\description{
Remove \code{data_list} from an existing data store.
}
