% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{rename_SO2}
\alias{rename_SO2}
\title{rename_SO2}
\usage{
rename_SO2(x, so2_map, is_awb = FALSE)
}
\arguments{
\item{x}{Data object (typically from pipeline)}

\item{so2_map}{TODO}

\item{is_awb}{TODO}
}
\value{
Data object with \code{Non.CO2} changed to SO2 name for SO2 data.
}
\description{
Rename SO2 to regional SO2. (TODO)
}
\author{
BBL May 2017
}
