% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-helpers.R
\name{repeat_add_columns}
\alias{repeat_add_columns}
\title{repeat_add_columns}
\usage{
repeat_add_columns(x, y)
}
\arguments{
\item{x}{Data frame (tibble) to repeat}

\item{y}{A copy of \code{x} is created for each row of this tibble}
}
\value{
A repeated \code{x} with columns from \code{y} added.
}
\description{
Repeat a data frame for each entry in a second, binding the columns together.
}
\details{
This corresponds to \code{repeat_and_add_vector} in the old data system.
}
\examples{
x <- tibble::tibble(x = 1:3)
y <- tibble::tibble(y = c(4, 5), z = c(6, 7))
repeat_add_columns(x, y)
}
\author{
BBL
}
