% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{replace_GLU}
\alias{replace_GLU}
\title{replace_GLU}
\usage{
replace_GLU(d, map = basin_to_country_mapping,
  GLU_pattern = "^GLU[0-9]{3}$")
}
\arguments{
\item{d}{A tibble with a column named "GLU"}

\item{map}{A tibble with columns \code{GLU_code} and \code{GLU_name}}

\item{GLU_pattern}{Regular expression string to identify the GLU codes}
}
\value{
A tibble with codes substituted for pattern, or vice versa, depending on the original
contents of the \code{GLU} column.
}
\description{
Replace GLU numerical codes with names, and vice versa
}
