% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{return_data}
\alias{return_data}
\title{return_data}
\usage{
return_data(...)
}
\arguments{
\item{...}{Objects to handle}
}
\value{
Object ready for insertion into the data system data structure.
}
\description{
Construct a data structure of objects (\code{...}) and return it.
Abstracts this away from chunk function code.
}

