% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{run_chunk}
\alias{run_chunk}
\title{run_chunk}
\usage{
run_chunk(chunk, all_data)
}
\arguments{
\item{chunk}{Chunk name, character}

\item{all_data}{Data required by chunk}
}
\value{
Data generated by chunk
}
\description{
run_chunk
}
\details{
Putting this \code{do.call} into this one-line
function lets us mock it when testing.
}
