% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_xml_conversion}
\alias{run_xml_conversion}
\title{run_xml_conversion}
\usage{
run_xml_conversion(dot)
}
\arguments{
\item{dot}{The current state of the pipeline started from \code{create_xml}.}
}
\description{
Run the CSV to XML conversion using the model interface tool.  This method
is should be the final call in a pipeline started with \code{\link{create_xml}}
and one or more calls to \code{\link{add_xml_data}}.
}

