% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{same_precursors_as}
\alias{same_precursors_as}
\title{same_precursors_as}
\usage{
same_precursors_as(x, y)
}
\arguments{
\item{x}{An object (source of precursors)}

\item{y}{Another object (that should get same precursors as \code{x})}
}
\value{
\code{x} with precursors set to those of \code{y}.
}
\description{
Copy precursors from one object to another.
}
