% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-chunk.R
\name{sample_sample_makedata}
\alias{sample_sample_makedata}
\title{gcam-sample_sample_makedata}
\usage{
sample_sample_makedata(all_data)
}
\arguments{
\item{all_data}{A named list, holding necessary input data}
}
\value{
A named list with all output data.
}
\description{
gcam-sample_sample_makedata
}
\author{
Author name
}

