% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_drake_target_name}
\alias{sanitize_drake_target_name}
\title{sanitize_drake_target_name}
\usage{
sanitize_drake_target_name(chunk_names)
}
\arguments{
\item{chunk_names}{A character vector of chunk names to clean}
}
\value{
A cleaned up \code{chunk_names} which can be used as
targets in drake.
}
\description{
Clean up potential target names so they do not include any
characters which drake does not allow.  At the moment this
just means replacing slashes, which we include in FILE or XML
chunk inputs / outputs, with a '.'
}
