% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_chunkdata}
\alias{save_chunkdata}
\title{save_chunkdata}
\usage{
save_chunkdata(chunkdata, write_inputs = FALSE)
}
\arguments{
\item{chunkdata}{Named list of tibbles (data frames) to write}

\item{write_inputs}{Write data that were read as inputs, not computed?
Write data produced by chunks to csv files.}
}
\description{
save_chunkdata
}

