% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{screen_forbidden}
\alias{screen_forbidden}
\title{Screen a function for use of functions forbidden by data system style guide.}
\usage{
screen_forbidden(fn)
}
\arguments{
\item{fn}{The function to be tested.  This is the actual function object, not
the name of the function.}
}
\value{
Nx2 Character matrix of flagged lines and the test that tripped them
(empty vector, if none)
}
\description{
Certain functions are forbidden by the dsr style guide from being used in
code chunks.  This function tests a function for calls to forbidden functions
and flags the offending lines.
}
\author{
RL 19 Apr 2017
}
