% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{set_subsector_shrwt}
\alias{set_subsector_shrwt}
\title{set_subsector_shrwt}
\usage{
set_subsector_shrwt(data, value_col = "calOutputValue")
}
\arguments{
\item{data}{Tibble to operate on}

\item{value_col}{Column with values to be used in setting subsector share-weights}
}
\value{
Tibble returned with a new column of calculated subsector shareweights.
}
\description{
Calculate subsector shareweights in calibration periods, where subsectors may have multiple technologies
}
