% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{set_traded_names}
\alias{set_traded_names}
\title{set_traded_names}
\usage{
set_traded_names(data, apply_selected_only = TRUE)
}
\arguments{
\item{data}{Data set to operate on}

\item{apply_selected_only}{Logical: only apply to region 1?}
}
\value{
Modified data
}
\description{
Convert names of traded secondary goods to be contained within region 1, with region appended to subsector and tech names
}
