% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{set_years}
\alias{set_years}
\title{set_years}
\usage{
set_years(data)
}
\arguments{
\item{data}{Tibble with text descriptions of model time periods to be replaced with numerical values.}
}
\value{
Modified tibble with 'numerical' values instead of text.
}
\description{
Replace text descriptions of years in exogenous input CSVs with numerical values. This allows model time periods
to be modified without requiring similar modifications in many input CSV tables.
}
\details{
Text strings include \code{start-year}, \code{final-calibration-year}, \code{final-historical-year},
\code{initial-future-year}, \code{initial-nonhistorical-year}, and \code{end-year}.
}
\note{
The returned 'numerical' values are actually characters; this helper function doesn't touch column types.
}
