% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeconomics.R
\name{socioeconomics_L100.GDP_hist_makedata}
\alias{socioeconomics_L100.GDP_hist_makedata}
\title{socioeconomics_L100.GDP_hist_makedata}
\usage{
socioeconomics_L100.GDP_hist_makedata(all_data)
}
\arguments{
\item{all_data}{A named list, holding all data system products so far}
}
\value{
A named list with all \code{socioeconomics_L100.GDP_hist} data.
}
\description{
socioeconomics_L100.GDP_hist_makedata
}

