% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-socioeconomics.R
\name{socioeconomics_L100.Population_downscale_ctry_makedata}
\alias{socioeconomics_L100.Population_downscale_ctry_makedata}
\title{socioeconomics_L100.Population_downscale_ctry_makedata}
\usage{
socioeconomics_L100.Population_downscale_ctry_makedata(all_data)
}
\arguments{
\item{all_data}{A named list, holding all data system products so far}
}
\value{
A named list with all \code{L100.Population_downscale_ctry} data.
}
\description{
socioeconomics_L100.Population_downscale_ctry_makedata
}

