% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-helpers.R
\name{standardize_iso}
\alias{standardize_iso}
\title{Standardize the ISO column by converting to lower case and renaming to 'iso'}
\usage{
standardize_iso(d, col = "iso", delete_original = TRUE)
}
\arguments{
\item{d}{The data to be transformed.}

\item{col}{The column currently containing the iso codes (default = 'iso')}

\item{delete_original}{Delete original column? Logical}
}
\value{
Tibble with the iso codes converted to lower case. If the column with
  the codes was not called 'iso', it will be renamed to 'iso'.
}
\description{
Most of the processing functions expect ISO codes to be in lower case and in
a column called ISO.
}
