% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{warn_data_injects}
\alias{warn_data_injects}
\title{warn_data_injects}
\usage{
warn_data_injects()
}
\value{
Number of temporary data objects being used inappropriately.
}
\description{
Check whether chunks are using any temporary (old data system data) 'injected' inputs.
}
