% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{write_to_all_regions}
\alias{write_to_all_regions}
\title{write_to_all_regions}
\usage{
write_to_all_regions(data, names, has.traded = FALSE, apply.to = "selected",
  set.market = FALSE)
}
\arguments{
\item{data}{Data}

\item{names}{Names}

\item{has.traded}{has.traded}

\item{apply.to}{apply.to}

\item{set.market}{set.market}
}
\value{
return
}
\description{
write_to_all_regions
}
