% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_data}
\alias{add_data}
\title{add_data}
\usage{
add_data(data_list, all_data)
}
\arguments{
\item{data_list}{List of data frames or other objects}

\item{all_data}{An existing (possibly empty) data store}
}
\value{
An empty data store
}
\description{
Add \code{data_list} to an existing data store.
}

