% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_dsflags}
\alias{add_dsflags}
\title{add_dsflag}
\usage{
add_dsflags(x, flags)
}
\arguments{
\item{x}{An object}

\item{comments}{A character vector of flags}
}
\value{
\code{x} with flags appended to any existing flags
}
\description{
Add character flags to a data system object. Flags are use internally, and in some
cases (for testing data) are written out with the data when the file is saved.
}

