% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{add_xml_data}
\alias{add_xml_data}
\title{Add a table to an XML pipeline to include for conversion to XML.}
\usage{
add_xml_data(dot, data, header, column_order_lookup = header)
}
\arguments{
\item{dot}{The current state of the pipeline started from \code{create_xml}}

\item{data}{The tibble of data to add to the conversion}

\item{header}{The header tag to can be looked up in the header file to
convert \code{data}}

\item{column_order_lookup}{A tag that can be used to look up \code{LEVEL2_DATA_NAMES}
to reorder the columns of data before XML conversion to ensure the correspond
with the ModelInterface header.  Note by default the \code{header} is used and if
given \code{NULL} no column reordering will be done.}
}
\value{
A "data structure" to hold the various parts needed to run the model
interface CSV to XML conversion.
}
\description{
We need the tibble to convert and a header tag which can be looked up in
the header file to convert the tibble.  This method is meant to be included
in a pipeline between calls of \code{\link{create_xml}} and \code{\link{run_xml_conversion}}.
Please see the wiki for more details \url{https://github.com/JGCRI/gcamdata/wiki/XML-Conversion}
}
\author{
PP March 2017
}
