% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{append_GLU}
\alias{append_GLU}
\title{append_GLU}
\usage{
append_GLU(data, ...)
}
\arguments{
\item{data}{Data, a tibble}

\item{...}{Names of variables to concatenate with \code{GLU} column, character}
}
\value{
A tibble with the \code{...} variable names concatenated with the \code{GLU}.
}
\description{
Append GLU to all specified variables
}
