% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{cmp_xml_files}
\alias{cmp_xml_files}
\title{Run the XML diff utility on two XML files}
\usage{
cmp_xml_files(fleft, fright, raw = FALSE)
}
\arguments{
\item{fleft}{The first file to compare}

\item{fright}{The second file to compare}

\item{raw}{Flag: if true run in raw mode}
}
\description{
Comparing two XML files turns out to be more easily done in python, so this
function runs the python comparison code.  The standard mode checks the
return code and returns TRUE if the files are equivalent, or FALSE if they
are not.  If the files are not equivalent, the differences found will *not*
be written to the console; you have to use raw mode if you want to get that
information.
}
\details{
The 'raw' mode returns the stdout generated by the comparison code as a
string. If the files are equivalent this return value will be empty (i.e., a
zero-length character vector, not an empty string).  If not, then it will
contain the diagnostic output. In either case the string returned will have a
'status' attribute giving the return code.
}
