% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{downscale_FAO_country}
\alias{downscale_FAO_country}
\title{downscale_FAO_country}
\usage{
downscale_FAO_country(data, country_name, dissolution_year,
  years = aglu.AGLU_HISTORICAL_YEARS)
}
\arguments{
\item{data}{Data to downscale, tibble}

\item{country_name}{Pre-dissolution country name, character}

\item{dissolution_year}{Year of country dissolution, integer}

\item{years}{Years to operate on, integer vector}
}
\value{
Downscaled data.
}
\description{
Helper function to downscale the countries that separated into
multiple modern countries (e.g. USSR).
}
