% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{driver}
\alias{driver}
\title{driver}
\usage{
driver(write_outputs = FALSE)
}
\arguments{
\item{write_outputs}{Write all chunk outputs to disk?}
}
\value{
a list of all built data
}
\description{
Run the entire data system
}

