% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{driver}
\alias{driver}
\title{driver}
\usage{
driver(all_data = empty_data(), write_outputs = TRUE, quiet = FALSE,
  stop_before = "", stop_after = "", outdir = OUTPUTS_DIR,
  xmldir = XML_DIR)
}
\arguments{
\item{all_data}{Data to be pre-loaded into data system}

\item{write_outputs}{Write all chunk outputs to disk?}

\item{quiet}{Suppress output?}

\item{stop_before}{Stop immediately before this chunk (character)}

\item{outdir}{Location to write output data (ignored if \code{write_outputs} is \code{FALSE})}

\item{xmldir}{Location to write output XML (ignored if \code{write_outputs} is \code{FALSE})}

\item{stop_before}{Stop immediately after this chunk  (character)}
}
\value{
A list of all built data.
}
\description{
Run the entire data system.
}
\details{
The driver loads any necessary data from input files,
runs all code chunks in an order dictated by their dependencies,
does error-checking, and writes outputs. For more details, see
the relevant wiki page at \url{ https://github.com/bpbond/gcamdata/wiki/Driver}.
}
\author{
BBL
}
