% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{driver}
\alias{driver}
\title{driver}
\usage{
driver(all_data = empty_data(), write_outputs = TRUE, quiet = FALSE)
}
\arguments{
\item{all_data}{Data to be pre-loaded into data system}

\item{write_outputs}{Write all chunk outputs to disk?}

\item{quiet}{Suppress output?}
}
\value{
A list of all built data.
}
\description{
Run the entire data system.
}
\details{
The driver loads any necessary data from input files,
runs all code chunks in an order dictated by their dependencies,
does error-checking, and writes outputs. For more details, see
the relevant wiki page at \url{ https://github.com/bpbond/gcamdata/wiki/Driver}.
}
\author{
BBL
}

