% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstrace.R
\name{dstrace}
\alias{dstrace}
\title{dstrace}
\usage{
dstrace(object_name, gcam_data_map = GCAM_DATA_MAP,
  previous_tracelist = NULL, recurse = TRUE)
}
\arguments{
\item{object_name}{Name of object to trace (can be either a data object or a code chunk)}

\item{gcam_data_map}{A tibble of metadata information; normally a built-in package dataset}

\item{previous_tracelist}{Information about previous objects printed}

\item{recurse}{Recurse to print information about precursor objects?}
}
\value{
A tibble with the trace information (object name and trace number)
}
\description{
dstrace
}
\examples{
dstrace("L100.FAO_ag_Exp_t")
}
\author{
BBL
}
