% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstrace.R
\name{dstrace}
\alias{dstrace}
\title{dstrace}
\usage{
dstrace(object_name, all_data, previous_tracelist = NULL, recurse = TRUE)
}
\arguments{
\item{object_name}{Name of object to trace (can be either a data object or a code chunk)}

\item{previous_tracelist}{Information about previous objects printed}

\item{recurse}{Recurse to print information about precursor objects?}
}
\value{
Boolean (object_name found, or not)
}
\description{
dstrace
}
\author{
BBL
}
