% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_header}
\alias{find_header}
\title{find_header}
\usage{
find_header(fqfn)
}
\arguments{
\item{fqfn}{Fully qualified filename, character}
}
\value{
The header as a character vector.
}
\description{
Read a file line-by-line to find how far its header extends, and return it.
}
\note{
Headers are defined as consecutive lines beginning with "#" at the top of a file.
}
\author{
Alexey Shiklomanov
}
