% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-helpers.R
\name{gdp_deflator}
\alias{gdp_deflator}
\title{gdp_deflator}
\source{
U.S. Bureau of Economic Analysis, Gross domestic product (implicit price deflator) [A191RD3A086NBEA], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/A191RD3A086NBEA, April 12, 2017
}
\usage{
gdp_deflator(year, base_year)
}
\arguments{
\item{year}{Year for which to calculate deflator}

\item{base_year}{Base year in deflator calculation}
}
\value{
Deflator (GDP in \code{from_year} divided by GDP in \code{to_year}).
}
\description{
Calculate a gross domestic product (GDP) implicit price deflator between two years.
}
\details{
The GDP deflator is a measure of price inflation/deflation with respect to a specific base year.
}
\note{
This returns slightly different values if \code{OLD_DATA_SYSTEM_BEHAVIOR} is TRUE.
}
\examples{
gdp_deflator(2010, base_year = 1990)
}
\author{
BBL
}
