% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data}
\alias{get_data}
\title{getdata}
\usage{
get_data(all_data, name)
}
\arguments{
\item{all_data}{Data structure}

\item{name}{Name of data to return}
}
\value{
Data structure (currently, a tibble or data frame)
}
\description{
This function returns a data frame (currently) in \code{all_data},
abstract away the mechanism for access it.
}

