% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{get_logit_fn_tables}
\alias{get_logit_fn_tables}
\title{get_logit_fn_tables}
\usage{
get_logit_fn_tables(data, names, base_header, GCAM_region_names,
  include_equiv_table = TRUE, write_all_regions = FALSE,
  default_logit_type = gcam.LOGIT_TYPES[1], ...)
}
\arguments{
\item{data}{Base data table indicating sectors and/or subsectors to be assigned a logit function type, tibble}

\item{names}{Column names to use in the data returned, character}

\item{base_header}{Base table name that is concatenated with the \code{logit.type} column to name each data table and header (ID string
for CSV to XML conversion), character}

\item{GCAM_region_names}{GCAM region names and ID numbers, tibble}

\item{include_equiv_table}{Logical indicating whether to generate an EQUIV_TABLE which will be called in the
CSVtoXML conversion, re-naming default logit type tags to the appropriate logit function type
in any other tables being read by the same batch XML file.}

\item{write_all_regions}{Logical indicating whether to call \code{\link{write_to_all_regions}}}

\item{default_logit_type}{Default logit function to use if the user did not specify one, character}

\item{...}{Other parameters to pass to \code{\link{write_to_all_regions}}}
}
\value{
A list, each element of which contains a data table and a header name to be used as the ID string for the
CSV to XML conversion.
}
\description{
Generate a list of tables that sets the appropriate logit choice function to use for supplysectors or subsectors.
}
\details{
The data table is partitioned into a list, with one element for each logit type. Each element of
the returned list has two sub-elements: 1) the header name (i.e., ID string) for the table, and 2) the data for the table.
If requested, an additional \code{EQUIV_TABLE} list will be included so that other data tables that contain the numerical
values of the logit exponents are assigned to the correct logit type. These tables with the numerical exponent values
are not required to be partitioned into separate tables according to logit type.
}
\note{
For documentation of the options for logit choice functions in GCAM, see \url{http://jgcri.github.io/gcam-doc/choice.html}.
The \code{EQUIV_TABLE} needs to be read once (and only once) per XML file created.
}
