% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime.R
\name{modeltime_declaredata}
\alias{modeltime_declaredata}
\title{modeltime_declaredata}
\usage{
modeltime_declaredata()
}
\value{
Names of all modeltime data
}
\description{
Declare what data are produced by \code{modeltime}
}

