% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA105.an_FAO_R_C_Y.R
\name{module_aglu_LA105.an_FAO_R_C_Y}
\alias{module_aglu_LA105.an_FAO_R_C_Y}
\title{module_aglu_LA105.an_FAO_R_C_Y}
\usage{
module_aglu_LA105.an_FAO_R_C_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L105.an_Food_Mt_R_C_Y}, \code{L105.an_Food_Pcal_R_C_Y}, \code{L105.an_kcalg_R_C_Y}, \code{L105.an_Prod_Mt_R_C_Y}, \code{L105.an_Prod_Mt_ctry_C_Y}, \code{L105.an_StockShares_R_BufGoat_2005}. The corresponding file in the
original data system was \code{LA105.an_FAO_R_C_Y.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
