% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA108.ag_Feed_R_C_Y.R
\name{module_aglu_LA108.ag_Feed_R_C_Y}
\alias{module_aglu_LA108.ag_Feed_R_C_Y}
\title{module_aglu_LA108.ag_Feed_R_C_Y}
\usage{
module_aglu_LA108.ag_Feed_R_C_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L108.ag_Feed_Mt_R_C_Y}, \code{L108.ag_NetExp_Mt_R_FodderHerb_Y}. The corresponding file in the
original data system was \code{LA108.ag_Feed_R_C_Y.R} (aglu level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
