% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB120.LC_GIS_R_LTgis_Yh_GLU.R
\name{module_aglu_LB120.LC_GIS_R_LTgis_Yh_GLU}
\alias{module_aglu_LB120.LC_GIS_R_LTgis_Yh_GLU}
\title{module_aglu_LB120.LC_GIS_R_LTgis_Yh_GLU}
\usage{
module_aglu_LB120.LC_GIS_R_LTgis_Yh_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L120.LC_bm2_R_LT_Yh_GLU}, \code{L120.LC_bm2_R_UrbanLand_Yh_GLU}, \code{L120.LC_bm2_R_Tundra_Yh_GLU}, \code{L120.LC_bm2_R_RckIceDsrt_Yh_GLU}, \code{L120.LC_bm2_ctry_LTsage_GLU}, \code{L120.LC_bm2_ctry_LTpast_GLU}. The corresponding file in the
original data system was \code{LB120.LC_GIS_R_LTgis_Yh_GLU.R} (aglu level1).
}
\description{
Land cover by GCAM region / aggregate land type / historical year / GLU.
}
\details{
Aggregate the \code{L100.Land_type_area_ha} dataset, interpolate land use historical
years, and split into various sub-categories. Missing values are set to zero because the GLU files don't include
zero values (i.e. they only report nonzero land use combinations).
}
\author{
BBL April 2017
}
