% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB122.LC_R_Cropland_Yh_GLU.R
\name{module_aglu_LB122.LC_R_Cropland_Yh_GLU}
\alias{module_aglu_LB122.LC_R_Cropland_Yh_GLU}
\title{module_aglu_LB122.LC_R_Cropland_Yh_GLU}
\usage{
module_aglu_LB122.LC_R_Cropland_Yh_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L122.ag_HA_to_CropLand_R_Y_GLU}, \code{L122.ag_EcYield_kgm2_R_C_Y_GLU}, \code{L122.LC_bm2_R_OtherArableLand_Yh_GLU}, \code{L122.LC_bm2_R_ExtraCropLand_Yh_GLU}, \code{L122.LC_bm2_R_HarvCropLand_C_Yh_GLU}, \code{L122.LC_bm2_R_HarvCropLand_Yh_GLU}. The corresponding file in the
original data system was \code{LB122.LC_R_Cropland_Yh_GLU.R} (aglu level1).
}
\description{
This chunk integrates disparate data sources for landcover and harvested area from FAO, Monfreda, and Hyde in order to calculate OtherArableLand
at the GCAM region-GLU level for each year, including historical years.
}
\details{
First, Hyde land cover data is aggregated to the region-GLU level in each year, and supplemented with Monfreda harvested area
information when necessary. Second, FAO fallowland, cropland, and harvested areadata is aggregated from the country level to the region
level. These processed data sources are used to calculate harvested area and available cropland area in each region-GLU in each year.
Finally, fallowland information at the region-GLU level is combined with the residual area = harvested area - available cropland area
at the region-GLU level to calculate OtherArableLand quantities, along with available ExtraCropLand quantities.
}
\author{
ACS April 2017
}
