% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L102.ghg_en_USA_S_T_Y.R
\name{module_emissions_L102.ghg_en_USA_S_T_Y}
\alias{module_emissions_L102.ghg_en_USA_S_T_Y}
\title{module_emissions_L102.ghg_en_USA_S_T_Y}
\usage{
module_emissions_L102.ghg_en_USA_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L102.ghg_tgej_USA_en_Sepa_F_2005}. The corresponding file in the
original data system was \code{L102.ghg_en_USA_S_T_Y.R} (emissions level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
HCM April 2017
}
