% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L125.bcoc_unmgd_R_S_T_Y.R
\name{module_emissions_L125.bcoc_unmgd_R_S_T_Y}
\alias{module_emissions_L125.bcoc_unmgd_R_S_T_Y}
\title{module_emissions_L125.bcoc_unmgd_R_S_T_Y}
\usage{
module_emissions_L125.bcoc_unmgd_R_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L125.bcoc_tgbkm2_R_grass_2000}, \code{L125.bcoc_tgbkm2_R_forest_2000}, \code{L125.deforest_coefs_bcoc}. The corresponding file in the
original data system was \code{L125.bcoc_unmgd_R_S_T_Y.R} (emissions level1).
}
\description{
Historical emissions from unmanaged land by land cover type, computed from RCP emissions data.
}
\details{
Calculate BC/OC emissions from unmanaged lands (savanna burning, forest fires, deforestation).
Downscale EDGAR regional emissions to GLU using shares of land area in each GLU within each region.
Divide forest-related emissions into deforestation and forest fires using GFED data
Compute global average deforestation emissions coefficients using deforestation from 2000 to 2005 using RCP emissions.
Note: Non-CO2s are calculated in a separate chunk.
}
\author{
RMH May 2017
}
