% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2112.ag_nonco2_IRR_MGMT.R
\name{module_emissions_L2112.ag_nonco2_IRR_MGMT}
\alias{module_emissions_L2112.ag_nonco2_IRR_MGMT}
\title{module_emissions_L2112.ag_nonco2_IRR_MGMT}
\usage{
module_emissions_L2112.ag_nonco2_IRR_MGMT(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{object}, \code{L2112.AWBEmissions}, \code{L2112.AGREmissions}. The corresponding file in the
original data system was \code{L2112.ag_nonco2_IRR_MGMT.R} (emissions level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}
