% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{module_energy_LA113.MSW}
\alias{module_energy_LA113.MSW}
\title{module_energy_LA113.MSW}
\usage{
module_energy_LA113.MSW(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}.
}
\description{
Construct the \code{energy} data structures.
}
\author{
BBL
}

