% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1322.Fert.R
\name{module_energy_LB1322.Fert}
\alias{module_energy_LB1322.Fert}
\title{module_energy_LB1322.Fert}
\usage{
module_energy_LB1322.Fert(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1322.Fert_Prod_MtN_R_F_Y}, \code{L1322.IO_R_Fert_F_Yh}, \code{L1322.in_EJ_R_indenergy_F_Yh}, \code{L1322.in_EJ_R_indfeed_F_Yh}, \code{L1322.Fert_NEcost_75USDkgN_F}. The corresponding file in the
original data system was \code{LB1322.Fert.R} (energy level1).
}
\description{
This chunk computes fertilizer production and energy inputs by technology.
}
\details{
Fertilizer production and energy information was calculated using country-level fuel share and energy intensity data
to allocate generic fertilizer production data across technologies. Re-allocation of energy use between energy and feedstock
quantities for the industrial sector was done to avoid negative values at the country level, before aggregating
to the regional level. Also, a final cost table was calculated using USA market fertilizer price data and H2A characteristics
of hydrogen production technologies.
}
\author{
AJS July 2017
}
