% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L254.transportation_USA.R
\name{module_gcam.usa_L254.transportation_USA}
\alias{module_gcam.usa_L254.transportation_USA}
\title{module_gcam.usa_L254.transportation_USA}
\usage{
module_gcam.usa_L254.transportation_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L254.DeleteSupplysector_USAtrn}, \code{L254.DeleteFinalDemand_USAtrn},
\code{L254.Supplysector_trn_USA}, \code{L254.FinalEnergyKeyword_trn_USA}, \code{L254.tranSubsectorLogit_USA},
\code{L254.tranSubsectorShrwtFllt_USA}, \code{L254.tranSubsectorInterp_USA}, \code{L254.tranSubsectorSpeed_USA},
\code{L254.tranSubsectorSpeed_passthru_USA}, \code{L254.tranSubsectorSpeed_noVOTT_USA},
\code{L254.tranSubsectorSpeed_nonmotor_USA}, \code{L254.tranSubsectorVOTT_USA}, \code{L254.tranSubsectorFuelPref_USA},
\code{L254.StubTranTech_USA}, \code{L254.StubTranTech_passthru_USA}, \code{L254.StubTranTech_nonmotor_USA},
\code{L254.StubTranTechLoadFactor_USA}, \code{L254.StubTranTechCost_USA}, \code{L254.StubTranTechCoef_USA},
\code{L254.PerCapitaBased_trn_USA}, \code{L254.PriceElasticity_trn_USA}, \code{L254.IncomeElasticity_trn_USA},
\code{L254.StubTranTechCalInput_USA}, \code{L254.StubTranTechProd_nonmotor_USA}, \code{L254.StubTranTechCalInput_passthru_USA},
\code{L254.BaseService_trn_USA}. The corresponding file in the
original data system was \code{L254.transportation_USA.R} (gcam-usa level2).
}
\description{
Generates GCAM-USA model inputs for transportation sector by states.
}
\details{
This chunk generates input files for transportation sector with generic information for supplysector,
subsector and technologies, as well as calibrated inputs and outputs by the US states.
}
\author{
RC Oct 2017
}
