% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_L100.GDP_hist.R
\name{module_socioeconomics_L100.GDP_hist}
\alias{module_socioeconomics_L100.GDP_hist}
\title{module_socioeconomics_L100.GDP_hist}
\usage{
module_socioeconomics_L100.GDP_hist(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.gdp_mil90usd_ctry_Yh}. The corresponding file in the
original data system was \code{L100.GDP_hist.R} (socioeconomics level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
BBL
}

