% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_L100.Population_downscale_ctry.R
\name{module_socioeconomics_L100.Population_downscale_ctry_DISABLED}
\alias{module_socioeconomics_L100.Population_downscale_ctry_DISABLED}
\title{module_socioeconomics_L100.Population_downscale_ctry}
\usage{
module_socioeconomics_L100.Population_downscale_ctry_DISABLED(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.Pop_thous_ctry_Yh}, \code{L100.Pop_thous_SSP_ctry_Yfut}. The corresponding file in the
original data system was \code{L100.Population_downscale_ctry.R} (socioeconomics level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
YourInitials CurrentMonthName 2017
}

