% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L201.Pop_GDP_scenarios.R
\name{module_socioeconomics_L201.Pop_GDP_scenarios}
\alias{module_socioeconomics_L201.Pop_GDP_scenarios}
\title{module_socioeconomics_L201.Pop_GDP_scenarios}
\usage{
module_socioeconomics_L201.Pop_GDP_scenarios(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L201.InterestRate}, \code{L201.Pop_GCAM3}, \code{L201.BaseGDP_GCAM3}, \code{L201.LaborForceFillout}, \code{L201.LaborProductivity_GCAM3}, \code{L201.PPPConvert}, \code{object}, \code{L201.BaseGDP_Scen}, \code{L201.LaborForceFillout}, \code{object2}, \code{L201.PPPConvert}. The corresponding file in the
original data system was \code{L201.Pop_GDP_scenarios.R} (socioeconomics level2).
}
\description{
Briefly describe what this chunk does.
}
\details{
Describe in detail what this chunk does.
}
\author{
HM&RH June 2017
}
