% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L145.water.demand.municipal.R
\name{module_water_L145.water.demand.municipal}
\alias{module_water_L145.water.demand.municipal}
\title{module_water_L145.water.demand.municipal}
\usage{
module_water_L145.water.demand.municipal(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L145.municipal_water_R_W_Yh_km3}, \code{L145.municipal_water_cost_R_75USD_m3}, \code{L145.municipal_water_eff_R_Y}. The corresponding file in the
original data system was \code{L145.water.demand.municipal.R} (water level1).
}
\description{
Generate municipal water withdrawals, municipal water base delivery cost, and municipal water use efficiency.
}
\details{
Describe in detail what this chunk does.
}
\author{
YL May 2017
}
