% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_chunkdata}
\alias{save_chunkdata}
\title{save_chunkdata}
\usage{
save_chunkdata(chunkdata)
}
\arguments{
\item{chunkdata}{Named list of data frames to write
Write data produced by chunks to csv files.}

\item{chunkdata}{List of tibbles to write}
}
\description{
save_chunkdata
}

