% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{set_traded_names}
\alias{set_traded_names}
\title{set_traded_names}
\usage{
set_traded_names(data, GCAM_region_names, apply_selected_only = TRUE)
}
\arguments{
\item{data}{Tibble to operate on}

\item{GCAM_region_names}{Tibble with GCAM region names and ID numbers}

\item{apply_selected_only}{Logical indicating whether region/subsector/technology re-assignment is applied to
the whole tibble, or only selected rows}
}
\value{
Tibble returned with modified region, subsector, and/or technology information.
}
\description{
Re-set region names in tables with traded secondary goods so that the traded secondary goods are all contained
within one specified region, with the(actual) region name prepended to the subsector and technology names(where applicable)
}
