% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{set_traded_names}
\alias{set_traded_names}
\title{set_traded_names}
\usage{
set_traded_names(data, GCAM_region_names, apply_selected_only = TRUE)
}
\arguments{
\item{data}{Data set to operate on, tibble}

\item{GCAM_region_names}{GCAM region names, ordered character vector}

\item{apply_selected_only}{Only apply to traded region \code{gcam.USA_CODE} (1)? Logical}
}
\value{
Modified data
}
\description{
Convert names of traded secondary goods to be contained within region 1, with region appended to subsector and tech names
}
