% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{set_years}
\alias{set_years}
\title{set_years}
\usage{
set_years(data)
}
\arguments{
\item{data}{Data with entries to be replaced, tibble}
}
\value{
Modified tibble with \code{start-year}, \code{final-calibration-year}, etc., converted to 'numeric' values
}
\description{
Replace text descriptions of years with numerical values
}
\note{
The new 'numeric' values are actually characters; this helper function doesn't touch column types.
}
