% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{tibbelize_outputs}
\alias{tibbelize_outputs}
\title{tibbelize_outputs}
\usage{
tibbelize_outputs(chunk_data, chunk_name)
}
\arguments{
\item{chunk_name}{Name of current chunk, character}

\item{chunk_outputs}{List of chunk outputs, a data list (see \link{\code{is_data_list}})}
}
\value{
A tibble with chunk name, output name, title, units, flags, precursors, and comments.
This table has one row per output name; multiple flags, precursors, etc., are concatenated into single entries.
}
\description{
Extract precursor and other metadata from chunk output data and convert to tibble form.
}
