% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{write_to_all_regions}
\alias{write_to_all_regions}
\title{write_to_all_regions}
\usage{
write_to_all_regions(data, names, has_traded = FALSE,
  apply_selected_only = TRUE, set_market = FALSE)
}
\arguments{
\item{data}{Data set to operate on, tibble}

\item{names}{Column names to return, character}

\item{apply_selected_only}{Only apply to region 1? Logical}

\item{set_market}{Overwrite \code{market} column with \code{region} data? Logical}

\item{has.traded}{has.traded TODO, logical}
}
\value{
TODO
}
\description{
write_to_all_regions
}
